+++ scripts/online_posterior_sampling.py
@@ -1,35 +1,48 @@
 #!/usr/bin/env python
 
 """
-Heuristic decision making
+Online Posterior Sampling for POMDP under STAR settings
 
+Papers:
+1. Posterior Sampling-based Online Learning for Episodic POMDPs, Tang, et al., 2024. https://doi.org/10.48550/arXiv.2310.10107
+2. (More) Efficient Reinforcement Learning via Posterior Sampling, Osband, et al., NIPS 2013. https://papers.nips.cc/paper_files/paper/2013/file/6a5889bb0190d0211a991f47bb19a777-Paper.pdf
 """
 import rospy
+import actionlib
+import json
+import math
 from time import process_time
 import pickle
 import numpy as np
-import actionlib
+from loss_fcns import *
 from pruning import *
+import project_utils as pu
 from nav_msgs.msg import Odometry
-from nav_msgs.srv import GetPlan
 from std_msgs.msg import Int8, Float32
 from move_base_msgs.msg import MoveBaseAction, MoveBaseGoal
+from nav_msgs.srv import GetPlan
+from int_preservation.srv import flevel, flevelRequest
 from status import areaStatus, battStatus, robotStatus
 from reset_simulation import *
-from heuristic_fcns import *
-from loss_fcns import *
-from int_preservation.srv import clusterAssignment, clusterAssignmentResponse
-from int_preservation.srv import areaAssignment
-from int_preservation.srv import assignmentAccomplishment
-from int_preservation.srv import registerRobot
-from int_preservation.srv import pauseSimulation
 
+def request_fmeasure(area, msg=True):
+    """
+    Service request for F-measure
+    :param msg:
+    :return: (tlapse, F) #tlapse here is time lapsed since decay model of area has been updated
+    """
+    rospy.wait_for_service("/flevel_server_" + str(area))
+    flevel_service = rospy.ServiceProxy("/flevel_server_" + str(area), flevel)
+    request = flevelRequest(msg)
+    result = flevel_service(request)
+    return result.fmeasure
 
 
 INDEX_FOR_X = 0
 INDEX_FOR_Y = 1
-SUCCEEDED = 3 #GoalStatus ID for succeeded, http://docs.ros.org/en/api/actionlib_msgs/html/msg/GoalStatus.html
+SUCCEEDED = 3  # GoalStatus ID for succeeded, http://docs.ros.org/en/api/actionlib_msgs/html/msg/GoalStatus.html
 SHUTDOWN_CODE = 99
+
 
 class Robot:
     def __init__(self, node_name):
@@ -43,150 +56,106 @@ class Robot:
 
         rospy.init_node(node_name, anonymous=True)
 
-        #Parameters
+        # Parameters
         self.robot_id = rospy.get_param("~robot_id")
         self.debug_mode = rospy.get_param("/debug_mode")
-        self.robot_velocity = rospy.get_param("/robot_velocity") #Linear velocity of robot; we assume linear and angular are relatively equal
-        self.gamma = rospy.get_param("/gamma") #discount factor
+        self.robot_velocity = rospy.get_param("/robot_velocity")  # Linear velocity of robot; we assume linear and angular are relatively equal
+        self.gamma = rospy.get_param("/gamma")  # discount factor
         self.max_fmeasure = rospy.get_param("/max_fmeasure")  # Max F-measure of an area
-        self.max_battery = rospy.get_param("/max_battery") #Max battery
-        self.battery_reserve = rospy.get_param("/battery_reserve") #Battery reserve
+        self.max_battery = rospy.get_param("/max_battery")  # Max battery
+        self.battery_reserve = rospy.get_param("/battery_reserve")  # Battery reserve
 
         f_thresh = rospy.get_param("/f_thresh")
-        self.fsafe, self.fcrit = f_thresh  # (safe, crit)
+        self.fsafe, self.fcrit = f_thresh
 
         batt_consumed_per_time = rospy.get_param("/batt_consumed_per_time")
-        self.batt_consumed_per_travel_time, self.batt_consumed_per_restored_f = batt_consumed_per_time  # (travel, restoration)
+        self.batt_consumed_per_travel_time, self.batt_consumed_per_restored_f = batt_consumed_per_time
 
-        self.dec_steps = rospy.get_param("/dec_steps")
+        self.dec_steps = rospy.get_param("/dec_steps")  # STAR
         self.restoration = rospy.get_param("/restoration")
         self.noise = rospy.get_param("/noise")
+        self.nareas = rospy.get_param("/nareas")  # Sample nodes from voronoi equal to area count #STAR
+        self.areas = [int(i + 1) for i in range(self.nareas)]  # list of int area IDs
         self.tolerance = rospy.get_param("/move_base_tolerance")
         self.t_operation = rospy.get_param("/t_operation")  # total duration of the operation
         self.save = rospy.get_param("/save")  # Whether to save data
-        self.task_scheduler = rospy.get_param("/task_scheduler") #task scheduler
 
-        #Initialize variables
-        self.init_x, self.init_y = rospy.get_param("~initial_pose_x"), rospy.get_param("~initial_pose_y") #Initialize robot pose
-        self.debug("Init x, y {}".format((self.init_x, self.init_y)))
-        charging_station_coords = self.init_x, self.init_y #rospy.get_param("/charging_station_coords")
+        # Initialize variables
+        charging_station_coords = rospy.get_param("~initial_pose_x"), rospy.get_param("~initial_pose_y")  # rospy.get_param("/charging_station_coords")
         charging_pose_stamped = pu.convert_coords_to_PoseStamped(charging_station_coords)
-        self.nodes_poses = [charging_pose_stamped] #list container for sampled nodes of type PoseStamped, where 0 is the charging station for that robot
+        self.sampled_nodes_poses = [charging_pose_stamped]  # list container for sampled nodes of type PoseStamped
 
-        #Pickle load the sampled area poses
+        # Pickle load the sampled area poses
         with open('{}.pkl'.format(rospy.get_param("/file_sampled_areas")), 'rb') as f:
             sampled_areas_coords = pickle.load(f)
-        nareas = rospy.get_param("/nareas")
-        for area_coords in sampled_areas_coords['n{}_p{}'.format(nareas, rospy.get_param("/placement"))]:
+        for area_coords in sampled_areas_coords['n{}_p{}'.format(self.nareas, rospy.get_param("/placement"))]:
             pose_stamped = pu.convert_coords_to_PoseStamped(area_coords)
-            self.nodes_poses.append(pose_stamped)
+            self.sampled_nodes_poses.append(pose_stamped)
 
-        self.x, self.y = self.init_x, self.init_y  # Initialize robot pose
+        self.x, self.y = 0.0, 0.0  # Initialize robot pose
         self.charging_station = 0
-        self.curr_loc_idx = self.charging_station #Initial location robot is the charging station
-        self.battery = self.max_battery #Initialize battery at max, then gets updated by subscribed battery topic
-        self.best_decision_idx = None
+        self.curr_loc = self.charging_station  # Initial location robot is the charging station
+        self.battery = self.max_battery  # Initialize battery at max, then gets updated by subscribed battery topic
+        self.best_decision = None
         self.dist_matrix = None
-        # self.sampled_nodes_poses = None
-        self.mission_area_idx = None
+        self.mission_area = None
         self.robot_status = robotStatus.IDLE.value
         self.available = True
+        self.curr_fmeasures = dict()  # container of current F-measure of areas
+        self.decay_rates_dict = dict()  # dictionary for decay rates
+        self.recorded_fdata = dict()  # dictionary of recorded data collected during mission per area
+        self.tlapses = dict()  # dictionary containing tlapse for each area since last restoration
 
-        self.state = list() #list of states
-        self.strict_bounds_list = list() #list of strict lower and upper bounds
+        for area in self.areas:
+            self.curr_fmeasures[area] = self.max_fmeasure
+            self.decay_rates_dict[area] = None
+            self.recorded_fdata[area] = list()
+            self.tlapses[area] = 0
 
-        self.decisions_made, self.decisions_accomplished, self.status_history = [], [], [] #record of data
-        self.total_dist_travelled = 0 #total distance travelled
-        self.process_time_counter = [] #container for time it took to come up with decision
-        self.assigned_areas = None
+        self.decay_rates_counter = 0  # counter for stored decay rates; should be equal to number of areas
+        self.decisions_made, self.decisions_accomplished, self.status_history = [], [], []  # record of data
+        self.total_dist_travelled = 0  # total distance travelled
+        self.process_time_counter = []  # container for time it took to come up with decision
 
-        self.requested_pause = False #indicator variable whether requested Stage to pause simulation
+        #Variables for online PSRL
+        self.inference = rospy.get_param("/inference") #Whether oracle or None; toggle CONSIDER_REPLAN which updates decay belief by posterior sampling
+        self.observations_dict = dict()
+        self.posterior_belief_disbn = dict()
+        init_mean, init_sd = 0.001875, 0.000875 #Prior data/mission (params are based on actual experimental settings)
+        for area in self.areas:
+            self.observations_dict[area] = list()
+            self.posterior_belief_disbn[area] = (init_mean, init_sd)
 
-        #TODO: Variables for online PSRL
+        # We sum this up
+        self.environment_status = dict()
+        for node in range(self.nareas + 1):
+            self.environment_status[node] = 999
 
-        #Publishers/Subscribers
+        # Publishers/Subscribers
         rospy.Subscriber('/robot_{}/odom'.format(self.robot_id), Odometry, self.distance_travelled_cb, queue_size=1)
 
         # Service request to move_base to get plan : make_Plan
         server = '/robot_' + str(self.robot_id) + '/move_base_node/make_plan'
         rospy.wait_for_service(server)
         self.get_plan_service = rospy.ServiceProxy(server, GetPlan)
-        self.debug("Getplan service: {}".format(self.get_plan_service))
+        # self.debug("Getplan service: {}".format(self.get_plan_service))
 
         rospy.Subscriber('/robot_{}/battery_status'.format(self.robot_id), Int8, self.battery_status_cb)
         rospy.Subscriber('/robot_{}/battery'.format(self.robot_id), Float32, self.battery_level_cb)
 
-        self.robot_status_pub = rospy.Publisher('/robot_{}/robot_status'.format(self.robot_id), Int8, queue_size=1)
-        self.mission_area_idx_pub = rospy.Publisher('/robot_{}/mission_area'.format(self.robot_id), Int8, queue_size=1)
-        self.location_pub = rospy.Publisher('/robot_{}/location'.format(self.robot_id), Int8, queue_size=1)
+        for area in self.areas:
+            rospy.Subscriber('/area_{}/decay_rate'.format(area), Float32, self.decay_rate_cb, area)
+            rospy.Subscriber('/area_{}/fmeasure'.format(area), Float32, self.area_fmeasure_cb, area)  # REMARK: Here we assume that we have live measurements of the F-measures
+            rospy.Subscriber('/area_{}/status'.format(area), Int8, self.area_status_cb, area)
 
-        #Action client to move_base
-        self.robot_goal_client = actionlib.SimpleActionClient('/robot_' + str(self.robot_id) + '/move_base', MoveBaseAction)
+        self.robot_status_pub = rospy.Publisher('/robot_{}/robot_status'.format(self.robot_id), Int8, queue_size=1)
+        self.mission_area_pub = rospy.Publisher('/robot_{}/mission_area'.format(self.robot_id), Int8, queue_size=1)
+
+        # Action client to move_base
+        self.robot_goal_client = actionlib.SimpleActionClient('/robot_' + str(self.robot_id) + '/move_base',
+                                                              MoveBaseAction)
         self.robot_goal_client.wait_for_server()
 
-        #Server for assigned cluster to monitor/preserve
-        self.cluster_assignment_server = rospy.Service("/cluster_assignment_server_" + str(self.robot_id), clusterAssignment, self.cluster_assignment_cb)
-
-        #TODO: Need to publish assignment_status for the re-assignment case
-
-        """
-        On charging:
-            Robot's mission area is 0. It then changes its status to CHARGING once it reaches the charging station.
-            The charging station, which subscribes to robot's status, charges up the battery.
-            Here, we assume there is only one charging station.
-            
-            If the robot_status is other than CHARGING, the battery status is DEPLETING.
-        
-        On area restoration:
-            Robot's mission area is a specific area. If reaches the area, it changes its status to RESTORING_F.
-            Now, the current mission area, which subscribes to both robot_status and robot_mission_area topics, will restore F; while,
-                those other areas not the mission area will have their F continually decay 
-        """
-
-    def register_to_central(self):
-        """
-        Register robot info to central
-        :return:
-        """
-        if self.task_scheduler == "central_planner":
-            rospy.wait_for_service('/robots_registry_server')
-            try:
-                register_request = rospy.ServiceProxy('/robots_registry_server', registerRobot)
-                self.debug("Registering robot info to central (id, x, y): {}, {}, {}".format(self.robot_id, self.init_x, self.init_y))
-                resp = register_request(self.robot_id, self.init_x, self.init_y)
-                self.debug("Registry to central success: {}".format(resp.registered))
-            except rospy.ServiceException as e:
-                rospy.logerr(f"Register to central service call failed: {e}")
-
-
-    def area_assignment_notice(self, assigned_areas):
-        """
-        Tags areas with the assigned robot
-        :param assigned_areas:
-        :return:
-        """
-        for area_id in assigned_areas:
-            rospy.wait_for_service("/area_assignment_server_" + str(area_id))
-            try:
-                area_assign = rospy.ServiceProxy("/area_assignment_server_" + str(area_id), areaAssignment)
-                resp = area_assign(self.robot_id)
-                self.debug("Noted assigned Area {}".format(resp))
-            except rospy.ServiceException as e:
-                rospy.logerr(f"Service call failed: {e}")
-
-    def cluster_assignment_cb(self, msg):
-        """
-        Sets the cluster assignment as the areas for restoration/preservation
-        :param msg:
-        :return:
-        """
-        self.robot_status = robotStatus.IDLE.value #Halts all operations of the robot to re-consider new assignment
-        assigned_areas, decay_rates, tlapses = msg.cluster, msg.decay_rates, msg.tlapses
-        assigned_areas, decay_rates, tlapses = list(assigned_areas), list(decay_rates), list(tlapses)
-        self.debug("Cluster assignment: {}, {}. Decay rates: {}. Tlapses: {}".format(type(assigned_areas), assigned_areas, decay_rates, tlapses))
-        self.instantiate_variables(assigned_areas=assigned_areas, decay_rates=decay_rates, tlapses=tlapses)
-        self.is_assigned = True
-        return clusterAssignmentResponse(self.is_assigned)
 
     # METHODS: Node poses and distance matrix
     def get_plan_request(self, start_pose, goal_pose, tolerance):
@@ -295,21 +264,30 @@ class Robot:
 
     def action_send_done_cb(self, state, result, robot_id):
         """
-
+        What happens after robot succeeded in fulfilling assigned action (recharge or restore area)
         :param msg:
         :return:
         """
         if state == SUCCEEDED:
-            self.curr_loc_idx = self.mission_area_idx
-            self.update_robot_status(robotStatus.RESTORING_F)
-
-            if self.mission_area_idx == self.charging_station:
+            self.debug("Arrived at mision_area: {}".format(self.mission_area))
+            self.curr_loc = self.mission_area
+            self.decisions_accomplished.append(self.mission_area)
+            self.best_decision = None
+            if self.mission_area == self.charging_station:
                 self.update_robot_status(robotStatus.CHARGING)
-            mission_area = self.get_assigned_area_id(self.mission_area_idx)
-            self.decisions_accomplished.append(mission_area)
-            self.best_decision_idx = None
+            else:
+                data = request_fmeasure(self.curr_loc)
+                measured_f = float(data)
+                self.recorded_fdata[self.curr_loc].append(measured_f)
+                observed_decay_rate = get_decay_rate(self.max_fmeasure, measured_f, self.tlapses[self.curr_loc])
+                self.observations_dict[self.curr_loc].append(observed_decay_rate)
 
-    def mean_duration_decay(self, duration_matrix, area_idx):
+                self.debug("Prior belief decay distribution: {}".format(self.posterior_belief_disbn[self.curr_loc]))
+                self.bayesian_belief_update(self.curr_loc)
+                self.debug("Observed decay rate: {}. Decay rate history: {}. Updated posterior decay distribution: {}".format(observed_decay_rate, self.observations_dict[self.curr_loc], self.posterior_belief_disbn[self.curr_loc]))
+                self.update_robot_status(robotStatus.RESTORING_F)
+
+    def mean_duration_decay(self, duration_matrix, area):
         """
         Measures the average duration that an area decays (when robot commits to another decision than restoring area)
         Note that each column of the duration matrix represents that the corresponding area is the decision that the robot is committing to.
@@ -318,7 +296,7 @@ class Robot:
         :param area:
         :return:
         """
-        submatrix = np.delete(duration_matrix, area_idx, axis=1)
+        submatrix = np.delete(duration_matrix, area, axis=1)
         return np.mean(submatrix)
 
     def estimate_battery_params(self, decision, curr_battery, curr_loc, fmeasures, noise):
@@ -328,13 +306,12 @@ class Robot:
         :return:
         """
         # Battery consumed travel and preserve area (if not charging station)
-        self.debug("Estimate battery params fmeasure, decision: {}, {}".format(fmeasures, decision))
-        battery_consumption = self.consume_battery(start_area_idx=curr_loc, next_area_idx=decision,
+        battery_consumption = self.consume_battery(start_area=curr_loc, next_area=decision,
                                                    curr_measure=fmeasures[decision],
                                                    noise=noise)
         # Battery consumed travel back to charging station
-        battery_consumption_backto_charging_station = self.consume_battery(start_area_idx=decision,
-                                                                           next_area_idx=self.charging_station,
+        battery_consumption_backto_charging_station = self.consume_battery(start_area=decision,
+                                                                           next_area=self.charging_station,
                                                                            curr_measure=None,
                                                                            noise=noise)
         total_battery_consumption = battery_consumption + battery_consumption_backto_charging_station
@@ -343,57 +320,45 @@ class Robot:
 
         return total_battery_consumption, feasible_battery
 
-    #TODO: Insert methods for online PSRL
-
-    def greedy_best_decision(self):
+    def compute_reward(self, fmeasure):
         """
-
+        Computes reward of action based on loss that has been saved up
+        :param area:
+        :param duration:
         :return:
         """
-        #Measure duration matrix
-        duration_matrix = self.dist_matrix/self.robot_velocity
+        reward = -loss_fcn(self.max_fmeasure, fmeasure)
+        return reward
 
-        #Measure the average duration an area decays
-        mean_duration_decay_dict = dict()
-        for area_idx in self.areas: #NOTE: The area here is not zero-indexed; rather starts at index 1
-            mean_duration_decay_dict[area_idx] = self.mean_duration_decay(duration_matrix, area_idx)
-
-        #Evaluate decision
+    def policy_max_value(self):
+        """
+        Policy that takes the feasible action that has the max value
+        :return:
+        """
+        # Evaluate feasible decisions
         decision_array = []
-        for decision_idx in self.areas: #decision is an area_idx
+        for decision in self.areas:
             # Battery consumption
-            battery_consumption, feasible_battery = self.estimate_battery_params(decision_idx, self.battery, self.curr_loc_idx,
+            battery_consumption, feasible_battery = self.estimate_battery_params(decision, self.battery, self.curr_loc,
                                                                                  self.curr_fmeasures, self.noise)
             # self.debug("Batt consumption: {}. Feasible batt: {}".format(battery_consumption, feasible_battery))
-            if not prune(self.battery, battery_consumption, self.battery_reserve) and decision_idx != self.curr_loc_idx:
-                #Immediate loss in i=1
-                duration = self.compute_duration(self.curr_loc_idx, decision_idx, self.curr_fmeasures[decision_idx], self.restoration, self.noise)
 
-                #TODO: It is here where we edit, we don't reset the f-measure of the area we visit
-                updated_fmeasures = self.adjust_fmeasures(self.curr_fmeasures.copy(), decision_idx, duration)  # F-measure of areas adjusted accordingly, i.e., consequence of decision
+            if not prune(self.battery, battery_consumption, self.battery_reserve) and decision != self.curr_loc:
+                #Compute the value based on belief decay rate`
+                duration = self.compute_duration(self.curr_loc, decision, self.curr_fmeasures[decision],
+                                                 self.restoration, self.noise)
+                updated_fmeasures = self.adjust_fmeasures(self.curr_fmeasures.copy(), decision, duration)  # F-measure of areas adjusted accordingly, i.e., consequence of decision
+                value = self.compute_reward(updated_fmeasures[decision])
+                decision_array.append((decision, value, feasible_battery))
 
-                #TODO: Po, add a dummy fmeasures for the immediate cost computation where the f-measure of the are is not reset
-                immediate_cost_decision = self.compute_opportunity_cost(updated_fmeasures) #immediate opportunity cost
-                # self.debug("Current F-measures: {}".format(self.curr_fmeasures))
-                # self.debug("Feasible decision: {}. Duration: {}. Updated F: {}. Immediate loss: {}".format(decision, duration, updated_fmeasures, immediate_cost_decision))
+        best_decision = self.charging_station
 
-                #Heuristic loss for i=2...k
-                forecasted_cost_decision = heuristic_cost_decision(updated_fmeasures, self.decay_rates_dict, (self.fsafe, self.fcrit),
-                                                         self.gamma, self.dec_steps, mean_duration_decay_dict) #forecasted opportunity cost
+        if len(decision_array) > 0:
+            best_decision = self.get_max_value(decision_array)
 
-                # self.debug("Discounted future losses through {} steps: {}".format(self.dec_steps, forecasted_cost_decision))
-                evaluated_cost_decision = immediate_cost_decision + forecasted_cost_decision #total forecasted opportunity cost
-                # self.debug("Appending: {}".format((decision, evaluated_cost_decision, feasible_battery)))
-                decision_array.append((decision_idx, evaluated_cost_decision, feasible_battery))
+        return best_decision
 
-        best_decision_idx = self.charging_station
-
-        if len(decision_array)>0:
-            best_decision_idx = self.get_best_decision(decision_array)
-
-        return best_decision_idx
-
-    def compute_duration(self, start_area_idx, next_area_idx, curr_measure, restoration, noise):
+    def compute_duration(self, start_area, next_area, curr_measure, restoration, noise):
         """
         Computes (time) duration of operation, which includes travelling distance plus restoration, if any
         :param distance:
@@ -403,42 +368,44 @@ class Robot:
         """
 
         # Travel distance
-        distance = self.dist_matrix[int(start_area_idx), int(next_area_idx)]
-        distance += noise * distance #distance + noise
+        distance = self.dist_matrix[int(start_area), int(next_area)]
+        distance += noise * distance  # distance + noise
         time = (distance / self.robot_velocity)
 
-        #If next area is not the charging station: the restoration is the f-measure; else, the restoration is the battery level
-        if next_area_idx != self.charging_station:
+        # If next area is not the charging station: the restoration is the f-measure; else, the restoration is the battery level
+        if next_area != self.charging_station:
             max_restore = self.max_fmeasure
         else:
             max_restore = self.max_battery
 
-        #Restoration time: If there is need for restoration
+        # Restoration time: If there is need for restoration
         if (curr_measure is not None) and (restoration is not None):
-            restore_time = (max_restore - curr_measure)/restoration
+            restore_time = (max_restore - curr_measure) / restoration
             time += restore_time
 
         return time
 
-    def consume_battery(self, start_area_idx, next_area_idx, curr_measure, noise):
+    def consume_battery(self, start_area, next_area, curr_measure, noise):
         """
-        Estimates battery consumption for the duration of the visit next_area from start_area.
+        Consumes curr_battery for the duration of the operation.
         This duration includes the distance plus F-measure restoration, if any
+        :param curr_battery:
+        :param duration:
+        :return:
         """
 
-        #Batt consumed in travel
-        distance = self.dist_matrix[int(start_area_idx), int(next_area_idx)]
+        # Batt consumed in travel
+        distance = self.dist_matrix[int(start_area), int(next_area)]
         distance += noise * distance
         travel_time = (distance / self.robot_velocity)
         battery_consumed = self.batt_consumed_per_travel_time * travel_time
 
-        #Batt consumed in area restoration
-        if next_area_idx != self.charging_station:
+        if next_area != self.charging_station:
             battery_consumed += self.batt_consumed_per_restored_f * (self.max_fmeasure - curr_measure)
 
         return battery_consumed
 
-    def adjust_fmeasures(self, fmeasures, visit_area_idx, duration):
+    def adjust_fmeasures(self, fmeasures, visit_area, duration):
         """
         Adjusts the F-measures of all areas in robot's mind. The visit area will be restored to max, while the other areas will decay for
         t duration. Note that the charging station is not part of the areas to monitor. And so, if the visit_area is the
@@ -448,24 +415,17 @@ class Robot:
         :param t:
         :return:
         """
-
-        """
-        Verify:
-            1. visit_area index - area_idx
-            2. fmeasure area key - area_idx
-            3. decay_rates_dict key - area_idx
-        """
-
-        for area_idx in self.areas:
-            if area_idx == visit_area_idx:
-                fmeasures[area_idx] = self.max_fmeasure #TODO: This may need to be fixed for the objective computation
+        for area in self.areas:
+            if area == visit_area:
+                fmeasures[area] = self.max_fmeasure
             else:
-                tlapse_decay = get_time_given_decay(self.max_fmeasure, fmeasures[area_idx], self.decay_rates_dict[area_idx]) + duration
-                fmeasures[area_idx] = decay(self.decay_rates_dict[area_idx], tlapse_decay, self.max_fmeasure)
+                tlapse_decay = get_time_given_decay(self.max_fmeasure, fmeasures[area],
+                                                    self.decay_rates_dict[area]) + duration
+                fmeasures[area] = decay(self.decay_rates_dict[area], tlapse_decay, self.max_fmeasure)
 
         return fmeasures
 
-    def compute_opportunity_cost(self, fmeasures):
+    def compute_net_loss(self, fmeasures):
         """
         Computes the net loss, (i.e., the sum of losses) of the fmeasures, which is a consequence of a decision
         Steps:
@@ -477,319 +437,155 @@ class Robot:
         netloss = compute_cost_fmeasures(fmeasures, self.fsafe, self.fcrit)
         return netloss
 
-    def get_best_decision(self, dec_arr):
+    def get_max_value(self, dec_arr):
         """
-        Returns the area index of the best decision in an array by sorting forecasted_loss ascendingly first then by remaining feasible battery.
+        Returns the best decision in an array by sorting values descendingly first then by remaining feasible battery.
         :param tree:
         :return:
         """
         # Sort the decisions: the cost is key while the value is decision
-        sorted_decisions = sorted(dec_arr, key = lambda x: (x[-2], -x[-1]))
-        # self.debug("Decisions sorted by cost: {}".format(sorted_decisions))
-        # self.debug("Best decision (branch info): {}".format(sorted_decisions[0]))
-        best_decision_idx = sorted_decisions[0][0] #pick the decision with least net loss and most available feasible battery
-
-        return best_decision_idx
+        sorted_decisions = sorted(dec_arr, key=lambda x: (-x[-2], -x[-1]))
+        self.debug("Decisions sorted by value, feasible batt: {}".format(sorted_decisions))
+        self.debug("Best decision (branch info): {}".format(sorted_decisions[0]))
+        best_decision = sorted_decisions[0][0]  # pick the decision with least net loss and most available feasible battery
+        return best_decision
 
     def update_tlapses_areas(self):
         """
         Lapses all time elapsed for each area
         :return:
         """
-        self.sim_t += 1
-        for area_idx in self.tlapses:
-            self.tlapses[area_idx] += 1
-        # self.debug("Sim t: {}. Time elapsed since last restored: {}".format(self.sim_t, self.tlapses))
+        for area in self.areas:
+            self.tlapses[area] += 1
         self.debug("Time elapsed since last restored: {}".format(self.tlapses))
 
-    def strict_bounds(self):
+    def bayesian_belief_update(self, area):
         """
-        Estimates the strict and upper bounds of the opportunity cost of the schedule
-        :return: strict upper and lower bounds
-        """
-        # Strict lower bound
-        duration_matrix = self.dist_matrix / self.robot_velocity
-        min_decay_rate = min(self.decay_rates_dict.values())
-        min_tlapse = min(self.tlapses.values())
-        min_duration = np.min(duration_matrix)
+        Performs bayesian update on belief decay distribution for area
 
-        max_decay_rate = max(self.decay_rates_dict.values())
-        max_tlapse = max(self.tlapses.values())
-        max_duration = np.max(duration_matrix)
+        Note: Design assumes we have a new observation for
 
-        strict_lower_bound = 0
-        tlapse = min_tlapse
-        for i in range(1, self.dec_steps + 1):
-            tlapse += min_duration
-            fmeasure = decay(min_decay_rate, tlapse, self.max_fmeasure)
-            discounted_loss = (self.gamma**(i - 1)) * loss_fcn(self.max_fmeasure, fmeasure)
-            strict_lower_bound += discounted_loss
-        strict_lower_bound *= (self.nareas - 1)
-
-        # Strict upper bound
-        strict_upper_bound = 0
-        tlapse = max_tlapse
-        for i in range(1, self.dec_steps + 1):
-            tlapse += i * max_duration
-            fmeasure = decay(max_decay_rate, tlapse, self.max_fmeasure)
-            discounted_loss = (self.gamma**(i - 1)) * loss_fcn(self.max_fmeasure, fmeasure)
-            strict_upper_bound += discounted_loss
-        strict_upper_bound *= (self.nareas - 1)
-
-        #Measure average time steps for plotting
-        mean_duration = np.mean(duration_matrix)
-        mean_tsteps = math.ceil(self.dec_steps * mean_duration)
-
-        return strict_lower_bound, strict_upper_bound, mean_tsteps
-
-    def extract_sampled_node_poses(self, assigned_areas):
-        """
-        Extracts the node poses of the assigned clusters into a list where index 0 is the charging station pose
-        :param cluster:
         :return:
         """
-        poses = list()
-        poses.append(self.nodes_poses[0])
-        for area_id in assigned_areas:
-            poses.append(self.nodes_poses[area_id])
-        return poses
+        #Prior
+        prior_mean, prior_sd = self.posterior_belief_disbn[area] #list of current posterior belief per area, now set as prior
 
-    """
-    NOTE: We use area index in computations containers (except sampled_nodes_poses), 
-        while we use assigned area ids in area subscriptions and action sending, as well as debugs
-        Area index (area_idx): 1, 2, 3, ...
-        Actual area id (area_id): The actual ones from the environment, wherein the assignment is a subset of
-        Note that for containers that are list in nature, we may index specific/corresponding area index as [area_idx-1]
-    """
-    def unsubscribe_previous_area_topics(self):
+        #Empirical mean and sd
+        n = len(self.observations_dict[area])
+        mean_data = np.mean(self.observations_dict[area]) #list of observations per area
+        sd_data = np.std(self.observations_dict[area], ddof=1)
+
+        #Bayesian udpate
+        sd_posterior_sq = 1 / (1/prior_sd**2 + n/sd_data**2)
+        mean_posterior = ((prior_mean / prior_sd**2) + ((n*mean_data) / sd_data**2)) * sd_posterior_sq
+        sd_posterior = np.sqrt(sd_posterior_sq)
+
+        #Update belief
+        self.posterior_belief_disbn[area] = (mean_posterior, sd_posterior)
+
+    def sample_posterior_belief(self, area):
         """
-        Unsubscribes to previously assigned areas topics:
-            > fmeasure
-            > status
+        Samples decay rate from current posterior belief on decay distribution
         :return:
         """
-        for area_id in self.subscribe_fmeasures:
-            if self.subscribe_fmeasures[area_id] is not None:
-                self.subscribe_fmeasures[area_id].unsubscribe()
-                self.debug("Unsubscribed from previously assigned Area {} fmeasure topic".format(area_id))
 
-        for area_id in self.subscribe_statuses:
-            if self.subscribe_statuses[area_id] is not None:
-                self.subscribe_statuses[area_id].unsubscribe()
-                self.debug("Unsubscribed from previously assigned Area {} status topic".format(area_id))
+        mean_belief, sd_belief = self.posterior_belief_disbn[area]
+        decay_rate = np.random.normal(loc=mean_belief, scale=sd_belief)
+        return decay_rate
 
-    def instantiate_variables(self, assigned_areas, decay_rates, tlapses):
-        """
-        Instantiates variables for every new assigned areas/cluster
-        :param self:
-        :param assigned_areas:
-        :return:
-        """
-        self.debug("Instantiating variables")
-
-        self.assigned_areas = assigned_areas  # Has the same list indexing with self.areas
-        self.nareas = len(self.assigned_areas)  # Sample nodes from voronoi equal to area count #STAR
-        self.areas = [int(i + 1) for i in range(self.nareas)]  # list of int area indexes, starting at index=1
-
-        # Send notice to areas about their robot assignment
-        self.area_assignment_notice(self.assigned_areas)
-
-        self.curr_fmeasures = dict()  # container of current F-measure of areas
-        self.decay_rates_dict = dict()  # dictionary for decay rates
-        self.tlapses = dict()  # dictionary containing tlapses of areas
-
-        for area_idx in self.areas:
-            self.decay_rates_dict[area_idx] = decay_rates[area_idx-1] #Instantiate provided decay rates
-            self.tlapses[area_idx] = tlapses[area_idx-1] #Instantiate provided tlapses
-
-        # Unsubscribe from previous area topics
-        # self.unsubscribe_previous_area_topics() TODO: Debug this one for the case with re-assignment
-        self.subscribe_fmeasures = dict()
-        self.subscribe_statuses = dict()
-
-        for area_idx in self.areas:
-            area_id = self.get_assigned_area_id(area_idx)
-            # Decay rates provided by central.
-            # rospy.Subscriber('/area_{}/decay_rate'.format(area_id), Float32, self.decay_rate_cb, area_id)
-            self.subscribe_fmeasures[area_id] = rospy.Subscriber('/area_{}/fmeasure'.format(area_id), Float32, self.area_fmeasure_cb, area_id)  # REMARK: Here we assume that we have live measurements of the F-measures
-            self.subscribe_statuses[area_id] = rospy.Subscriber('/area_{}/status'.format(area_id), Int8, self.area_status_cb, area_id)
-
-        #Wait for self.curr_fmeasures to be populated
-        while set(list(self.curr_fmeasures.keys())) != set(self.areas):
-            self.debug("Subscribing to area topics...")
-            rospy.sleep(1)
-
-        # Sampled node poses and distance matrix
-        self.sampled_nodes_poses = self.extract_sampled_node_poses(self.assigned_areas)
-        self.build_dist_matrix()
-
-    def get_assigned_area_index(self, area_id):
-        """
-        Gets area index of assigned area, (note area index starts at 1)
-        :param area_id:
-        :return:
-        """
-        area_idx = self.assigned_areas.index(area_id) + 1
-        return area_idx
-
-    def get_assigned_area_id(self, area_idx):
-        """
-        Gets actual assigned area id given area index, (note area index starts at 1)
-        :param self:
-        :param area_idx:
-        :return:
-        """
-        area_id = self.assigned_areas[area_idx-1]
-        return area_id
-
-    def get_current_state(self):
-        """
-        Retrieves current state
-        :return:
-        """
-        tlapses = dict()
-        decay_rates = dict()
-        for area_idx in self.areas:
-            area_id = self.get_assigned_area_id(area_idx)
-            tlapses[area_id] = self.tlapses[area_idx]
-            decay_rates[area_id] = self.decay_rates_dict[area_idx]
-
-        # state = (self.sim_t, self.get_assigned_area_id(self.curr_loc_idx), self.battery, tlapses, decay_rates)
-        state = (self.get_assigned_area_id(self.curr_loc_idx), self.battery, tlapses, decay_rates)
-        return state
-
-    #TODO: Adapt online PSRL
-    #Methods: Run operation
     def run_operation(self, filename, freq=1):
         """
         :return:
         """
 
-        if self.robot_id < 999: #Not a dummy robot
+        if self.robot_id < 999:
             rate = rospy.Rate(freq)
-            rospy.sleep(5)  # Wait for nodes to register
 
-            self.register_to_central()
+            while len(self.sampled_nodes_poses) != self.nareas + 1:
+                self.debug("Insufficient data. Sampled nodes poses: {}/{}".format(len(self.sampled_nodes_poses), self.nareas + 1))
+                rate.sleep()  # Data for decay rates haven't registered yet
 
-            while self.dist_matrix is None:
-                self.debug("Initialization: No cluster assignment yet. Waiting for assignment...")
-                rospy.sleep(1)
+            for area in self.areas:
+                belief_decay = self.sample_posterior_belief(area) #Sample posterior belief based on prior knowledge
+                self.decay_rates_dict = belief_decay
+                self.observations_dict[area].append(belief_decay) #Append to history of observed decay, (set as initial)
 
+            self.debug("Sufficent data. Decay rates: {}. Sampled nodes poses: {}".format(self.decay_rates_dict,
+                                                                                         self.sampled_nodes_poses))  # Prior knowledge of decay rates
 
-            self.sim_t = 0
-            # while not rospy.is_shutdown() and self.sim_t<self.t_operation:
-            while not rospy.is_shutdown():
-                curr_state = self.get_current_state()
-                self.state.append(curr_state)
-                self.debug("Curr state: {}".format(curr_state))
+            self.build_dist_matrix()
+            t = 0
+            while not rospy.is_shutdown() and t < self.t_operation:
                 self.robot_status_pub.publish(self.robot_status)
-                self.location_pub.publish(self.get_assigned_area_id(self.curr_loc_idx))
                 self.status_history.append(self.robot_status)
-
-                #I am learning something new! Each day
-                # Freedom comes with courage. And with courage comes freedom.
-                # New wave. I create a new culture. A new trend people will appreciate and follow and love
-                # A brand new consciousness
-
-                if self.robot_status == robotStatus.IDLE.value: # Here, robot is available but unassigned
+                if self.robot_status == robotStatus.IDLE.value:
                     self.debug('Robot idle')
-                    self.debug("With dist_matrix: {}".format(bool(self.dist_matrix is not None)))
                     if self.dist_matrix is not None:
                         self.update_robot_status(robotStatus.READY)
 
                 elif self.robot_status == robotStatus.READY.value:
                     self.debug('Robot ready')
-
-                    #TODO: Insert pause request here
-                    self.request_pause(True) #Request pause simulation
-
                     think_start = process_time()
                     self.think_decisions()
                     think_end = process_time()
                     think_elapsed = self.time_elapsed(think_start, think_end)
-
-                    bounds = self.strict_bounds()
-                    self.debug('Strict bounds:{}'.format(bounds))
-                    self.strict_bounds_list.append(bounds)
-
                     self.process_time_counter.append(think_elapsed)
-
-                    self.debug('Best decision: {}. Process time: {}s'.format(self.get_assigned_area_id(self.best_decision_idx), think_elapsed))
-
-                    if self.requested_pause:
-                        self.request_pause(False)
-
+                    self.debug('Best decision: {}. Process time: {}s'.format(self.best_decision, think_elapsed))
                     self.update_robot_status(robotStatus.IN_MISSION)
 
                 elif self.robot_status == robotStatus.IN_MISSION.value:
                     self.debug('Robot in mission. Total distance travelled: {}'.format(self.total_dist_travelled))
-                    if self.available: #TODO: Insert the variable self.assigned for re-assignment case
+                    if self.available:
                         self.commence_mission()
 
-                #TODO: Insert CONSIDER_REPLAN
-
                 elif self.robot_status == robotStatus.CHARGING.value:
-                    #TODO: Po, we insert self.assign_status = unassigned here for central's reconsideration. HERE!
                     self.debug('Waiting for battery to charge up')
 
                 elif self.robot_status == robotStatus.RESTORING_F.value:
                     self.debug('Restoring F-measure')
 
-                # Update tlapse for each area when all nodes have registered
-                if len(self.decisions_made) > 1 or (self.robot_status != robotStatus.IDLE.value) and (self.robot_status != robotStatus.READY.value):
-                    self.update_tlapses_areas() #TODO: We can potentially insert here the tlapse from rospy.get_time()
+                elif self.robot_status == robotStatus.CONSIDER_REPLAN.value:
+                    self.debug('Consider re-plan...')
+                    #Posterior sampling of decay rates based on belief
+                    for area in self.areas:
+                        belief_decay = self.sample_posterior_belief(area)
+                        self.decay_rates_dict = belief_decay
+                    self.debug("Sampled posterior decay rates: {}".format(self.decay_rates_dict))
+                    self.update_robot_status(robotStatus.IN_MISSION)
 
-                if self.save:
-                    pu.dump_data(self.strict_bounds_list, '{}_strict_bounds'.format(filename))
-                    pu.dump_data(self.state, '{}_environment_state'.format(filename))
-                    pu.dump_data(self.process_time_counter, '{}_robot{}_process_time'.format(filename, self.robot_id))
-                    pu.dump_data(self.decisions_made, '{}_robot{}_decisions'.format(filename, self.robot_id))
-                    pu.dump_data((self.decisions_accomplished, self.total_dist_travelled), '{}_robot{}_decisions_acc_travel'.format(filename, self.robot_id))
-                    pu.dump_data(self.status_history, '{}_robot{}_status_history'.format(filename, self.robot_id))
-
+                if len(self.decisions_made)>0 or (self.robot_status != robotStatus.IDLE.value) and (
+                        self.robot_status != robotStatus.READY.value) and (
+                        self.robot_status != robotStatus.CONSIDER_REPLAN.value):
+                    self.update_tlapses_areas()  # Update the tlapse per area
+                    self.compute_curr_fmeasures()
+                t += 1
                 rate.sleep()
 
-            # #Store results
-            # self.update_robot_status(robotStatus.SHUTDOWN) #TODO: There should be a notice that come from central actually
-            # self.robot_status_pub.publish(self.robot_status)
-            # self.location_pub.publish(self.get_assigned_area_id(self.curr_loc_idx))
-            # self.status_history.append(self.robot_status)
-            #
-            # #Wait before all other nodes have finished dumping their data
-            # if self.save:
-            #     pu.dump_data(self.strict_bounds_list, '{}_strict_bounds'.format(filename))
-            #     pu.dump_data(self.state, '{}_environment_state'.format(filename))
-            #     pu.dump_data(self.process_time_counter, '{}_robot{}_process_time'.format(filename, self.robot_id))
-            #     pu.dump_data(self.decisions_made, '{}_robot{}_decisions'.format(filename, self.robot_id))
-            #     pu.dump_data((self.decisions_accomplished, self.total_dist_travelled), '{}_robot{}_decisions_acc_travel'.format(filename, self.robot_id))
-            #     pu.dump_data(self.status_history, '{}_robot{}_status_history'.format(filename, self.robot_id))
-            #     self.debug("Dumped all data.".format(self.robot_id))
-            # self.shutdown(sleep=10)
+            # Store results
+            self.update_robot_status(robotStatus.SHUTDOWN)
+            self.robot_status_pub.publish(self.robot_status)
+            self.status_history.append(self.robot_status)
 
-    """
-    NOTE: For the subscribed/published topics and data storages, as well as self.debugs! use assigned area ids! All else area index
-    """
-    def request_pause(self, is_pause=True):
-        """
-        Sends pause request to pause_simulation
-        :return:
-        """
-        self.requested_pause = is_pause
-        rospy.wait_for_service('/pause_queue_server')
-        try:
-            pause_request = rospy.ServiceProxy('/pause_queue_server', pauseSimulation)
-            agent_id = self.robot_id
-            resp = pause_request(is_pause, agent_id)
-            return resp.pause_result
-        except rospy.ServiceException as e:
-            rospy.logerr(f"Pause service call failed: {e}")
+            # Wait before all other nodes have finished dumping their data
+            if self.save:
+                if self.inference is not None:
+                    pu.dump_data((self.recorded_fdata, self.observations_dict),
+                                 '{}_robot{}_recorded_data'.format(filename, self.robot_id))
+
+                pu.dump_data(self.process_time_counter, '{}_robot{}_process_time'.format(filename, self.robot_id))
+                pu.dump_data(self.decisions_made, '{}_robot{}_decisions'.format(filename, self.robot_id))
+                pu.dump_data((self.decisions_accomplished, self.total_dist_travelled),
+                             '{}_robot{}_decisions_acc_travel'.format(filename, self.robot_id))
+                pu.dump_data(self.status_history, '{}_robot{}_status_history'.format(filename, self.robot_id))
+                self.debug("Dumped all data.".format(self.robot_id))
+            self.shutdown(sleep=10)
 
     def think_decisions(self):
         """
         Thinks of the best decision before starting mission
         :return:
         """
-        self.best_decision_idx = self.greedy_best_decision()
+        self.best_decision = self.policy_max_value()
 
     def time_elapsed(self, think_start, think_end):
         """
@@ -809,18 +605,15 @@ class Robot:
 
     def send2_next_area(self):
         """
-        Sends the robot to the next area in the optimal path:
+        Sends the robot to the next area in the optimal path/decision:
         :return:
         """
-        if self.best_decision_idx is not None:
-            self.mission_area_idx = self.best_decision_idx #TODO: Will the robot charge up battery?
-            mission_area_id = self.charging_station
-            if self.mission_area_idx != self.charging_station:
-                mission_area_id = self.get_assigned_area_id(self.mission_area_idx) #TODO: Will the robot charge up battery?
-            self.mission_area_idx_pub.publish(mission_area_id)
-            self.debug('Heading to: {}. {}'.format(mission_area_id, self.sampled_nodes_poses[self.mission_area_idx]))
-            self.decisions_made.append(mission_area_id) #store decisions made
-            self.go_to_target(self.mission_area_idx)
+        if self.best_decision is not None:
+            self.mission_area = self.best_decision
+            self.mission_area_pub.publish(self.mission_area)
+            self.debug('Heading to: {}. {}'.format(self.mission_area, self.sampled_nodes_poses[self.mission_area]))
+            self.decisions_made.append(self.mission_area)  # store decisions made
+            self.go_to_target(self.mission_area)
             return 1
         return 0
 
@@ -833,8 +626,8 @@ class Robot:
         self.robot_status = status.value
 
     def distance_travelled_cb(self, msg):
-        #Updates total distance travelled
-        #Sets curr robot pose
+        # Updates total distance travelled
+        # Sets curr robot pose
         x, y = msg.pose.pose.position.x, msg.pose.pose.position.y
         self.total_dist_travelled += math.dist((self.x, self.y), (x, y))
         self.x, self.y = x, y
@@ -853,25 +646,11 @@ class Robot:
         :param msg:
         :return:
         """
+        self.environment_status[self.charging_station] = msg.data
         if msg.data == battStatus.FULLY_CHARGED.value:
             if self.robot_id < 999: self.debug("Fully charged!")
             self.available = True
             self.update_robot_status(robotStatus.IN_MISSION)
-
-    def notify_assignment_accomplishment(self, area_id):
-        """
-        Assigns clusters to robots
-        :param clusters:
-        :return:
-        """
-
-        rospy.wait_for_service("/assignment_accomplishment_server")
-        try:
-            notify_server = rospy.ServiceProxy("/assignment_accomplishment_server", assignmentAccomplishment)
-            resp = notify_server(self.robot_id, area_id)
-            self.debug("Robot: {}. Accomplished assignment: {}. Notified server: {}".format(self.robot_id, area_id, resp.confirmation))
-        except rospy.ServiceException as e:
-            rospy.logerr(f"Service call failed: {e}")
 
     def area_status_cb(self, msg, area_id):
         """
@@ -879,16 +658,16 @@ class Robot:
         :param msg:
         :return:
         """
-        area_idx = self.get_assigned_area_index(area_id)
-        status = msg.data
-
+        self.environment_status[area_id] = msg.data
         if msg.data == areaStatus.RESTORED_F.value:
-            if self.robot_id < 999: self.debug("Area {} fully restored!".format(area_id))
-            self.tlapses[area_idx] = 0  # Reset the tlapse since last restored for the newly restored area
-            self.debug("Notifying server for accomplishment of restoring Area {}".format(area_id))
-            self.notify_assignment_accomplishment(area_id) #Notifies central that recent assignment is accomplished
             self.available = True
-            self.update_robot_status(robotStatus.IN_MISSION)
+            self.tlapses[area_id] = 0
+            if self.robot_id < 999: self.debug("Area {} fully restored! tlapse reset...".format(area_id))
+
+            if (self.inference is not None) and (self.inference != 'oracle'):
+                self.update_robot_status(robotStatus.CONSIDER_REPLAN)
+            else:
+                self.update_robot_status(robotStatus.IN_MISSION)
 
     def decay_rate_cb(self, msg, area_id):
         """
@@ -897,10 +676,10 @@ class Robot:
         :param area_id:
         :return:
         """
-        area_idx = self.get_assigned_area_index(area_id)
-        if self.assigned_areas is not None and self.decay_rates_dict[area_idx] == None:
-            if self.robot_id < 999: self.debug("Area {} decay rate: {}".format(area_id, msg.data))
-            self.decay_rates_dict[area_idx] = msg.data
+        if self.inference == 'oracle':
+            if self.decay_rates_dict[area_id] != msg.data: self.debug(
+                "Oracle knowledge, change in decay in area {}: {}".format(area_id, msg.data))
+            self.decay_rates_dict[area_id] = msg.data  # A subscribed topic. Oracle knows exactly the decay rate happening in area
 
     def area_fmeasure_cb(self, msg, area_id):
         """
@@ -909,8 +688,18 @@ class Robot:
         :param area_id:
         :return:
         """
-        area_idx = self.get_assigned_area_index(area_id)
-        self.curr_fmeasures[area_idx] = msg.data
+        if self.inference == 'oracle':
+            self.curr_fmeasures[area_id] = msg.data
+
+    def compute_curr_fmeasures(self):
+        """
+        Computes current fmeasures based on tlapse and decay rates
+        :return:
+        """
+        for area in self.areas:
+           self.curr_fmeasures[area] = decay(self.decay_rates_dict[area], self.tlapses[area], self.max_fmeasure)
+        # self.debug("Used for computation. Tlapses: {}. Decay rates: {}".format(self.tlapses, self.decay_rates_dict))
+        # self.debug("Computed current f-measures: {}".format(self.curr_fmeasures))
 
     def debug(self, msg):
         pu.log_msg('robot', self.robot_id, msg, self.debug_mode)
@@ -919,910 +708,8 @@ class Robot:
         self.debug("Reached {} time operation. Shutting down...".format(self.t_operation))
         kill_nodes(sleep)
 
+
 if __name__ == '__main__':
-    # os.chdir('/home/ameldocena/.ros/int_preservation/results')
     os.chdir('/root/catkin_ws/src/results/int_preservation')
     filename = rospy.get_param('/file_data_dump')
     Robot('online_posterior_sampling').run_operation(filename)
-
-
-#### Version prior to re-structuring
-
-# class Robot:
-#     def __init__(self, node_name):
-#         """
-#
-#         :param node_name:
-#         :param areas:
-#         :param est_distance_matrix:
-#         :param est_batt_consumption_matrix:
-#         """
-#
-#         rospy.init_node(node_name, anonymous=True)
-#
-#         # Parameters
-#         self.robot_id = rospy.get_param("~robot_id")
-#         self.debug_mode = rospy.get_param("/debug_mode")
-#         self.robot_velocity = rospy.get_param(
-#             "/robot_velocity")  # Linear velocity of robot; we assume linear and angular are relatively equal
-#         self.gamma = rospy.get_param("/gamma")  # discount factor
-#         self.max_fmeasure = rospy.get_param("/max_fmeasure")  # Max F-measure of an area
-#         self.max_battery = rospy.get_param("/max_battery")  # Max battery
-#         self.battery_reserve = rospy.get_param("/battery_reserve")  # Battery reserve
-#
-#         f_thresh = rospy.get_param("/f_thresh")
-#         self.fsafe, self.fcrit = f_thresh  # (safe, crit)
-#
-#         batt_consumed_per_time = rospy.get_param("/batt_consumed_per_time")
-#         self.batt_consumed_per_travel_time, self.batt_consumed_per_restored_f = batt_consumed_per_time  # (travel, restoration)
-#
-#         self.dec_steps = rospy.get_param("/dec_steps")
-#         self.restoration = rospy.get_param("/restoration")
-#         self.noise = rospy.get_param("/noise")
-#         self.tolerance = rospy.get_param("/move_base_tolerance")
-#         self.t_operation = rospy.get_param("/t_operation")  # total duration of the operation
-#         self.save = rospy.get_param("/save")  # Whether to save data
-#         self.task_scheduler = rospy.get_param("/task_scheduler")  # task scheduler
-#
-#         # Initialize variables
-#         self.init_x, self.init_y = rospy.get_param("~initial_pose_x"), rospy.get_param(
-#             "~initial_pose_y")  # Initialize robot pose
-#         self.debug("Init x, y {}".format((self.init_x, self.init_y)))
-#         charging_station_coords = self.init_x, self.init_y  # rospy.get_param("/charging_station_coords")
-#         charging_pose_stamped = pu.convert_coords_to_PoseStamped(charging_station_coords)
-#         self.nodes_poses = [
-#             charging_pose_stamped]  # list container for sampled nodes of type PoseStamped, where 0 is the charging station for that robot
-#
-#         # Pickle load the sampled area poses
-#         with open('{}.pkl'.format(rospy.get_param("/file_sampled_areas")), 'rb') as f:
-#             sampled_areas_coords = pickle.load(f)
-#         nareas = rospy.get_param("/nareas")
-#         for area_coords in sampled_areas_coords['n{}_p{}'.format(nareas, rospy.get_param("/placement"))]:
-#             pose_stamped = pu.convert_coords_to_PoseStamped(area_coords)
-#             self.nodes_poses.append(pose_stamped)
-#
-#         self.x, self.y = self.init_x, self.init_y  # Initialize robot pose
-#         self.charging_station = 0
-#         self.curr_loc_idx = self.charging_station  # Initial location robot is the charging station
-#         self.battery = self.max_battery  # Initialize battery at max, then gets updated by subscribed battery topic
-#         self.best_decision_idx = None
-#         self.dist_matrix = None
-#         self.sampled_nodes_poses = None
-#         self.mission_area_idx = None
-#         self.robot_status = robotStatus.IDLE.value
-#         self.available = True
-#
-#         self.state = list()  # list of states
-#         self.strict_bounds_list = list()  # list of strict lower and upper bounds
-#         self.nareas = None  # number of assigned areas will be supplied
-#
-#         self.decisions_made, self.decisions_accomplished, self.status_history = [], [], []  # record of data
-#         self.total_dist_travelled = 0  # total distance travelled
-#         self.process_time_counter = []  # container for time it took to come up with decision
-#         self.assigned_areas = None
-#
-#         self.environment_status = dict()
-#         self.environment_status[self.charging_station] = 999
-#
-#         # Publishers/Subscribers
-#         rospy.Subscriber('/robot_{}/odom'.format(self.robot_id), Odometry, self.distance_travelled_cb, queue_size=1)
-#
-#         # Service request to move_base to get plan : make_Plan
-#         server = '/robot_' + str(self.robot_id) + '/move_base_node/make_plan'
-#         rospy.wait_for_service(server)
-#         self.get_plan_service = rospy.ServiceProxy(server, GetPlan)
-#         self.debug("Getplan service: {}".format(self.get_plan_service))
-#
-#         rospy.Subscriber('/robot_{}/battery_status'.format(self.robot_id), Int8, self.battery_status_cb)
-#         rospy.Subscriber('/robot_{}/battery'.format(self.robot_id), Float32, self.battery_level_cb)
-#
-#         self.robot_status_pub = rospy.Publisher('/robot_{}/robot_status'.format(self.robot_id), Int8, queue_size=1)
-#         self.mission_area_idx_pub = rospy.Publisher('/robot_{}/mission_area'.format(self.robot_id), Int8, queue_size=1)
-#
-#         self.location_pub = rospy.Publisher('/robot_{}/location'.format(self.robot_id), Int8, queue_size=1)
-#
-#         # Action client to move_base
-#         self.robot_goal_client = actionlib.SimpleActionClient('/robot_' + str(self.robot_id) + '/move_base',
-#                                                               MoveBaseAction)
-#         self.robot_goal_client.wait_for_server()
-#
-#         # Server for assigned cluster to monitor/preserve
-#         self.cluster_assignment_server = rospy.Service("/cluster_assignment_server_" + str(self.robot_id),
-#                                                        clusterAssignment, self.cluster_assignment_cb)
-#
-#         """
-#         On charging:
-#             Robot's mission area is 0. It then changes its status to CHARGING once it reaches the charging station.
-#             The charging station, which subscribes to robot's status, charges up the battery.
-#             Here, we assume there is only one charging station.
-#
-#             If the robot_status is other than CHARGING, the battery status is DEPLETING.
-#
-#         On area restoration:
-#             Robot's mission area is a specific area. If reaches the area, it changes its status to RESTORING_F.
-#             Now, the current mission area, which subscribes to both robot_status and robot_mission_area topics, will restore F; while,
-#                 those other areas not the mission area will have their F continually decay
-#         """
-#
-#     def register_to_central(self):
-#         """
-#         Register robot info to central
-#         :return:
-#         """
-#         if self.task_scheduler == "central_planner":
-#             rospy.wait_for_service('/robots_registry_server')  # TODO
-#             try:
-#                 register_request = rospy.ServiceProxy('/robots_registry_server', registerRobot)  # TODO
-#                 self.debug("Registering robot info to central (id, x, y): {}, {}, {}".format(self.robot_id, self.init_x,
-#                                                                                              self.init_y))
-#                 resp = register_request(self.robot_id, self.init_x, self.init_y)  # TODO
-#                 self.debug("Registry to central success: {}".format(resp.registered))  # TODO
-#             except rospy.ServiceException as e:
-#                 rospy.logerr(f"Register to central service call failed: {e}")
-#
-#     def area_assignment_notice(self, assigned_areas):
-#         """
-#         Tags areas with the assigned robot
-#         :param assigned_areas:
-#         :return:
-#         """
-#         for area_id in assigned_areas:
-#             rospy.wait_for_service("/area_assignment_server_" + str(area_id))
-#             try:
-#                 area_assign = rospy.ServiceProxy("/area_assignment_server_" + str(area_id), areaAssignment)
-#                 resp = area_assign(self.robot_id)
-#                 self.debug("Noted assigned Area {}".format(resp))
-#             except rospy.ServiceException as e:
-#                 rospy.logerr(f"Service call failed: {e}")
-#
-#     def cluster_assignment_cb(self, msg):
-#         """
-#         Sets the cluster assignment as the areas for restoration/preservation
-#         :param msg:
-#         :return:
-#         """
-#         self.robot_status = robotStatus.IDLE.value  # Halts all operations of the robot to re-consider new assignment
-#         assigned_areas, decay_rates, tlapses = msg.cluster, msg.decay_rates, msg.tlapses
-#         assigned_areas, decay_rates, tlapses = list(assigned_areas), list(decay_rates), list(tlapses)
-#         self.debug(
-#             "Cluster assignment: {}, {}. Decay rates: {}. Tlapses: {}".format(type(assigned_areas), assigned_areas,
-#                                                                               decay_rates, tlapses))
-#         self.instantiate_variables(assigned_areas=assigned_areas, decay_rates=decay_rates, tlapses=tlapses)
-#         self.is_assigned = True
-#         return clusterAssignmentResponse(self.is_assigned)
-#
-#     # METHODS: Node poses and distance matrix
-#     def get_plan_request(self, start_pose, goal_pose, tolerance):
-#         """
-#         Sends a request to GetPlan service to create a plan for path from start to goal without actually moving the robot
-#         :param start_pose:
-#         :param goal_pose:
-#         :param tolerance:
-#         :return:
-#         """
-#         req = GetPlan()
-#         req.start = start_pose
-#         req.goal = goal_pose
-#         req.tolerance = tolerance
-#         server = self.get_plan_service
-#         result = server(req.start, req.goal, req.tolerance)
-#         path = result.plan.poses
-#         return path
-#
-#     def decouple_path_poses(self, path):
-#         """
-#         Decouples a path of PoseStamped poses; returning a list of x,y poses
-#         :param path: list of PoseStamped
-#         :return:
-#         """
-#         list_poses = list()
-#         for p in path:
-#             x, y = p.pose.position.x, p.pose.position.y
-#             list_poses.append((x, y))
-#         return list_poses
-#
-#     def compute_path_total_dist(self, list_poses):
-#         """
-#         Computes the total path distance
-#         :return:
-#         """
-#         total_dist = 0
-#         for i in range(len(list_poses) - 1):
-#             dist = math.dist(list_poses[i], list_poses[i + 1])
-#             total_dist += dist
-#         return total_dist
-#
-#     def compute_dist_bet_areas(self, area_i, area_j, tolerance):
-#         """
-#         Computes the distance between area_i and area_j:
-#             1. Call the path planner between area_i and area_j
-#             2. Decouple the elements of path planning
-#             3. Compute the distance then total distance
-#         :param area_i: PoseStamped
-#         :param area_j: PoseStamped
-#         :return:
-#         """
-#         path = self.get_plan_request(area_i, area_j, tolerance)
-#         list_poses = self.decouple_path_poses(path)
-#         total_dist = self.compute_path_total_dist(list_poses)
-#         return total_dist
-#
-#     def build_dist_matrix(self):
-#         """
-#         Builds the distance matrix among areas
-#         :return:
-#         """
-#         n = len(self.sampled_nodes_poses)
-#         self.dist_matrix = np.zeros((n, n))
-#
-#         for i in range(n):
-#             for j in range(n):
-#                 area_i, area_j = self.sampled_nodes_poses[i], self.sampled_nodes_poses[j]
-#                 if area_i != area_j:
-#                     dist = self.compute_dist_bet_areas(area_i, area_j, self.tolerance)
-#                     self.dist_matrix[i, j] = dist
-#
-#         self.debug("Dist matrix: {}".format(self.dist_matrix))
-#
-#     # METHODS: Send robot to area
-#     def go_to_target(self, goal_idx):
-#         """
-#         Action client to move_base to move to target goal
-#         Goal is PoseStamped msg
-#         :param: goal_idx, index of goal in sampled_nodes_poses list
-#         :return:
-#         """
-#         goal = self.sampled_nodes_poses[goal_idx]
-#         self.send_robot_goal(goal)
-#
-#     def send_robot_goal(self, goal):
-#         """
-#         Sends robot to goal via action client
-#         :param robot:
-#         :param goal: PoseStamped object
-#         :return:
-#         """
-#         movebase_goal = MoveBaseGoal()
-#         movebase_goal.target_pose = goal
-#         self.available = False
-#         action_goal_cb = (lambda state, result: self.action_send_done_cb(state, result, self.robot_id))
-#         self.robot_goal_client.send_goal(movebase_goal, done_cb=action_goal_cb, active_cb=self.action_send_active_cb)
-#
-#     def action_send_active_cb(self):
-#         """
-#         Sets robot as unavailable when pursuing goal
-#         :return:
-#         """
-#         self.available = False
-#         self.update_robot_status(robotStatus.IN_MISSION)
-#
-#     def action_send_done_cb(self, state, result, robot_id):
-#         """
-#
-#         :param msg:
-#         :return:
-#         """
-#         if state == SUCCEEDED:
-#             self.curr_loc_idx = self.mission_area_idx
-#             self.update_robot_status(robotStatus.RESTORING_F)
-#
-#             if self.mission_area_idx == self.charging_station:
-#                 self.update_robot_status(robotStatus.CHARGING)
-#             mission_area = self.get_assigned_area_id(self.mission_area_idx)
-#             self.decisions_accomplished.append(mission_area)
-#             self.best_decision_idx = None
-#
-#     def mean_duration_decay(self, duration_matrix, area_idx):
-#         """
-#         Measures the average duration that an area decays (when robot commits to another decision than restoring area)
-#         Note that each column of the duration matrix represents that the corresponding area is the decision that the robot is committing to.
-#         We thus delete the column of the area, and then take the average to measure the average duration.
-#         :param duration_matrix:
-#         :param area:
-#         :return:
-#         """
-#         submatrix = np.delete(duration_matrix, area_idx, axis=1)
-#         return np.mean(submatrix)
-#
-#     def estimate_battery_params(self, decision, curr_battery, curr_loc, fmeasures, noise):
-#         """
-#         Measures battery consumption and feasible battery
-#         :param decision:
-#         :return:
-#         """
-#         # Battery consumed travel and preserve area (if not charging station)
-#         self.debug("Estimate battery params fmeasure, decision: {}, {}".format(fmeasures, decision))
-#         battery_consumption = self.consume_battery(start_area_idx=curr_loc, next_area_idx=decision,
-#                                                    curr_measure=fmeasures[decision],
-#                                                    noise=noise)
-#         # Battery consumed travel back to charging station
-#         battery_consumption_backto_charging_station = self.consume_battery(start_area_idx=decision,
-#                                                                            next_area_idx=self.charging_station,
-#                                                                            curr_measure=None,
-#                                                                            noise=noise)
-#         total_battery_consumption = battery_consumption + battery_consumption_backto_charging_station
-#         # Feasible batt = current batt - (consumption of decision + consumption to travel back to charging station)
-#         feasible_battery = curr_battery - total_battery_consumption
-#
-#         return total_battery_consumption, feasible_battery
-#
-#     def greedy_best_decision(self):
-#         """
-#
-#         :return:
-#         """
-#         # Measure duration matrix
-#         duration_matrix = self.dist_matrix / self.robot_velocity
-#
-#         # Measure the average duration an area decays
-#         mean_duration_decay_dict = dict()
-#         for area_idx in self.areas:  # NOTE: The area here is not zero-indexed; rather starts at index 1
-#             mean_duration_decay_dict[area_idx] = self.mean_duration_decay(duration_matrix, area_idx)
-#
-#         # Evaluate decision
-#         decision_array = []
-#         for decision_idx in self.areas:  # decision is an area_idx
-#             # Battery consumption
-#             battery_consumption, feasible_battery = self.estimate_battery_params(decision_idx, self.battery,
-#                                                                                  self.curr_loc_idx,
-#                                                                                  self.curr_fmeasures, self.noise)
-#             # self.debug("Batt consumption: {}. Feasible batt: {}".format(battery_consumption, feasible_battery))
-#             if not prune(self.battery, battery_consumption, self.battery_reserve) and decision_idx != self.curr_loc_idx:
-#                 # Immediate loss in i=1
-#                 duration = self.compute_duration(self.curr_loc_idx, decision_idx, self.curr_fmeasures[decision_idx],
-#                                                  self.restoration, self.noise)
-#
-#                 # TODO: It is here where we edit, we don't reset the f-measure of the area we visit
-#                 updated_fmeasures = self.adjust_fmeasures(self.curr_fmeasures.copy(), decision_idx,
-#                                                           duration)  # F-measure of areas adjusted accordingly, i.e., consequence of decision
-#
-#                 # TODO: Po, add a dummy fmeasures for the immediate cost computation where the f-measure of the are is not reset
-#                 immediate_cost_decision = self.compute_opportunity_cost(updated_fmeasures)  # immediate opportunity cost
-#                 # self.debug("Current F-measures: {}".format(self.curr_fmeasures))
-#                 # self.debug("Feasible decision: {}. Duration: {}. Updated F: {}. Immediate loss: {}".format(decision, duration, updated_fmeasures, immediate_cost_decision))
-#
-#                 # Heuristic loss for i=2...k
-#                 forecasted_cost_decision = heuristic_cost_decision(updated_fmeasures, self.decay_rates_dict,
-#                                                                    (self.fsafe, self.fcrit),
-#                                                                    self.gamma, self.dec_steps,
-#                                                                    mean_duration_decay_dict)  # forecasted opportunity cost
-#
-#                 # self.debug("Discounted future losses through {} steps: {}".format(self.dec_steps, forecasted_cost_decision))
-#                 evaluated_cost_decision = immediate_cost_decision + forecasted_cost_decision  # total forecasted opportunity cost
-#                 # self.debug("Appending: {}".format((decision, evaluated_cost_decision, feasible_battery)))
-#                 decision_array.append((decision_idx, evaluated_cost_decision, feasible_battery))
-#
-#         best_decision_idx = self.charging_station
-#
-#         if len(decision_array) > 0:
-#             best_decision_idx = self.get_best_decision(decision_array)
-#
-#         return best_decision_idx
-#
-#     def compute_duration(self, start_area_idx, next_area_idx, curr_measure, restoration, noise):
-#         """
-#         Computes (time) duration of operation, which includes travelling distance plus restoration, if any
-#         :param distance:
-#         :param restoration: restore a measure (if not None) back to full measure per second
-#         :param noise: expected noise in distance travelled
-#         :return:
-#         """
-#
-#         # Travel distance
-#         distance = self.dist_matrix[int(start_area_idx), int(next_area_idx)]
-#         distance += noise * distance  # distance + noise
-#         time = (distance / self.robot_velocity)
-#
-#         # If next area is not the charging station: the restoration is the f-measure; else, the restoration is the battery level
-#         if next_area_idx != self.charging_station:
-#             max_restore = self.max_fmeasure
-#         else:
-#             max_restore = self.max_battery
-#
-#         # Restoration time: If there is need for restoration
-#         if (curr_measure is not None) and (restoration is not None):
-#             restore_time = (max_restore - curr_measure) / restoration
-#             time += restore_time
-#
-#         return time
-#
-#     def consume_battery(self, start_area_idx, next_area_idx, curr_measure, noise):
-#         """
-#         Estimates battery consumption for the duration of the visit next_area from start_area.
-#         This duration includes the distance plus F-measure restoration, if any
-#         """
-#
-#         # Batt consumed in travel
-#         distance = self.dist_matrix[int(start_area_idx), int(next_area_idx)]
-#         distance += noise * distance
-#         travel_time = (distance / self.robot_velocity)
-#         battery_consumed = self.batt_consumed_per_travel_time * travel_time
-#
-#         # Batt consumed in area restoration
-#         if next_area_idx != self.charging_station:
-#             battery_consumed += self.batt_consumed_per_restored_f * (self.max_fmeasure - curr_measure)
-#
-#         return battery_consumed
-#
-#     def adjust_fmeasures(self, fmeasures, visit_area_idx, duration):
-#         """
-#         Adjusts the F-measures of all areas in robot's mind. The visit area will be restored to max, while the other areas will decay for
-#         t duration. Note that the charging station is not part of the areas to monitor. And so, if the visit_area is the
-#         charging station, then all of the areas will decay as duration passes by.
-#         :param fmeasures:
-#         :param visit_area:
-#         :param t:
-#         :return:
-#         """
-#
-#         """
-#         Verify:
-#             1. visit_area index - area_idx
-#             2. fmeasure area key - area_idx
-#             3. decay_rates_dict key - area_idx
-#         """
-#
-#         for area_idx in self.areas:
-#             if area_idx == visit_area_idx:
-#                 fmeasures[area_idx] = self.max_fmeasure  # TODO: This may need to be fixed for the objective computation
-#             else:
-#                 tlapse_decay = get_time_given_decay(self.max_fmeasure, fmeasures[area_idx],
-#                                                     self.decay_rates_dict[area_idx]) + duration
-#                 fmeasures[area_idx] = decay(self.decay_rates_dict[area_idx], tlapse_decay, self.max_fmeasure)
-#
-#         return fmeasures
-#
-#     def compute_opportunity_cost(self, fmeasures):
-#         """
-#         Computes the net loss, (i.e., the sum of losses) of the fmeasures, which is a consequence of a decision
-#         Steps:
-#             1. Computes the loss for each of the F-measure of the areas
-#             2. Sums up the losses to get the cost of the decision
-#         :param fmeasures:
-#         :return:
-#         """
-#         netloss = compute_cost_fmeasures(fmeasures, self.fsafe, self.fcrit)
-#         return netloss
-#
-#     def get_best_decision(self, dec_arr):
-#         """
-#         Returns the area index of the best decision in an array by sorting forecasted_loss ascendingly first then by remaining feasible battery.
-#         :param tree:
-#         :return:
-#         """
-#         # Sort the decisions: the cost is key while the value is decision
-#         sorted_decisions = sorted(dec_arr, key=lambda x: (x[-2], -x[-1]))
-#         # self.debug("Decisions sorted by cost: {}".format(sorted_decisions))
-#         # self.debug("Best decision (branch info): {}".format(sorted_decisions[0]))
-#         best_decision_idx = sorted_decisions[0][
-#             0]  # pick the decision with least net loss and most available feasible battery
-#
-#         return best_decision_idx
-#
-#     def update_tlapses_areas(self, sim_t):
-#         """
-#         Lapses all time elapsed for each area
-#         :return:
-#         """
-#         # TODO: I think the sim_t should likewise be elapsed only when we elapsed time on the areas. This is within simulation
-#
-#         for area_idx in self.tlapses:
-#             self.tlapses[area_idx] += 1
-#         self.debug("Sim t: {}. Time elapsed since last restored: {}".format(sim_t, self.tlapses))
-#
-#     def strict_bounds(self):
-#         """
-#         Estimates the strict and upper bounds of the opportunity cost of the schedule
-#         :return: strict upper and lower bounds
-#         """
-#         # Strict lower bound
-#         duration_matrix = self.dist_matrix / self.robot_velocity
-#         min_decay_rate = min(self.decay_rates_dict.values())
-#         min_tlapse = min(self.tlapses.values())
-#         min_duration = np.min(duration_matrix)
-#
-#         max_decay_rate = max(self.decay_rates_dict.values())
-#         max_tlapse = max(self.tlapses.values())
-#         max_duration = np.max(duration_matrix)
-#
-#         strict_lower_bound = 0
-#         tlapse = min_tlapse
-#         for i in range(1, self.dec_steps + 1):
-#             tlapse += min_duration
-#             fmeasure = decay(min_decay_rate, tlapse, self.max_fmeasure)
-#             discounted_loss = (self.gamma ** (i - 1)) * loss_fcn(self.max_fmeasure, fmeasure)
-#             strict_lower_bound += discounted_loss
-#         strict_lower_bound *= (self.nareas - 1)
-#
-#         # Strict upper bound
-#         strict_upper_bound = 0
-#         tlapse = max_tlapse
-#         for i in range(1, self.dec_steps + 1):
-#             tlapse += i * max_duration
-#             fmeasure = decay(max_decay_rate, tlapse, self.max_fmeasure)
-#             discounted_loss = (self.gamma ** (i - 1)) * loss_fcn(self.max_fmeasure, fmeasure)
-#             strict_upper_bound += discounted_loss
-#         strict_upper_bound *= (self.nareas - 1)
-#
-#         # Measure average time steps for plotting
-#         mean_duration = np.mean(duration_matrix)
-#         mean_tsteps = math.ceil(self.dec_steps * mean_duration)
-#
-#         return strict_lower_bound, strict_upper_bound, mean_tsteps
-#
-#     def extract_sampled_node_poses(self, assigned_areas):
-#         """
-#         Extracts the node poses of the assigned clusters into a list where index 0 is the charging station pose
-#         :param cluster:
-#         :return:
-#         """
-#         poses = list()
-#         poses.append(self.nodes_poses[0])
-#         for area_id in assigned_areas:
-#             poses.append(self.nodes_poses[area_id])
-#         return poses
-#
-#     """
-#     NOTE: We use area index in computations containers (except sampled_nodes_poses),
-#         while we use assigned area ids in area subscriptions and action sending, as well as debugs
-#         Area index (area_idx): 1, 2, 3, ...
-#         Actual area id (area_id): The actual ones from the environment, wherein the assignment is a subset of
-#         Note that for containers that are list in nature, we may index specific/corresponding area index as [area_idx-1]
-#     """
-#
-#     def unsubscribe_previous_area_topics(self):
-#         """
-#         Unsubscribes to previously assigned areas topics:
-#             > fmeasure
-#             > status
-#         :return:
-#         """
-#         for area_id in self.subscribe_fmeasures:
-#             if self.subscribe_fmeasures[area_id] is not None:
-#                 self.subscribe_fmeasures[area_id].unsubscribe()
-#                 self.debug("Unsubscribed from previously assigned Area {} fmeasure topic".format(area_id))
-#
-#         for area_id in self.subscribe_statuses:
-#             if self.subscribe_statuses[area_id] is not None:
-#                 self.subscribe_statuses[area_id].unsubscribe()
-#                 self.debug("Unsubscribed from previously assigned Area {} status topic".format(area_id))
-#
-#     def instantiate_variables(self, assigned_areas, decay_rates, tlapses):
-#         """
-#         Instantiates variables for every new assigned areas/cluster
-#         :param self:
-#         :param assigned_areas:
-#         :return:
-#         """
-#         self.debug("Instantiating variables")
-#
-#         self.assigned_areas = assigned_areas  # Has the same list indexing with self.areas
-#         self.nareas = len(self.assigned_areas)  # Sample nodes from voronoi equal to area count #STAR
-#         self.areas = [int(i + 1) for i in range(self.nareas)]  # list of int area indexes, starting at index=1
-#
-#         # Send notice to areas about their robot assignment
-#         self.area_assignment_notice(self.assigned_areas)
-#
-#         self.curr_fmeasures = dict()  # container of current F-measure of areas
-#         self.decay_rates_dict = dict()  # dictionary for decay rates
-#         self.tlapses = dict()  # dictionary containing tlapses of areas
-#
-#         for area_idx in self.areas:
-#             self.decay_rates_dict[area_idx] = decay_rates[area_idx - 1]  # Instantiate provided decay rates
-#             self.tlapses[area_idx] = tlapses[area_idx - 1]  # Instantiate provided tlapses
-#
-#         # Environment status intialization
-#         environment_status = dict()
-#         environment_status[self.charging_station] = self.environment_status[self.charging_station]
-#         for node in range(self.nareas):
-#             self.environment_status[node + 1] = 999
-#         self.environment_status = environment_status
-#
-#         # Unsubscribe from previous area topics
-#         # self.unsubscribe_previous_area_topics() TODO: Debug this one
-#         self.subscribe_fmeasures = dict()
-#         self.subscribe_statuses = dict()
-#
-#         for area_idx in self.areas:
-#             area_id = self.get_assigned_area_id(area_idx)
-#             # Decay rates provided by central.
-#             # rospy.Subscriber('/area_{}/decay_rate'.format(area_id), Float32, self.decay_rate_cb, area_id)
-#             self.subscribe_fmeasures[area_id] = rospy.Subscriber('/area_{}/fmeasure'.format(area_id), Float32,
-#                                                                  self.area_fmeasure_cb,
-#                                                                  area_id)  # REMARK: Here we assume that we have live measurements of the F-measures
-#             self.subscribe_statuses[area_id] = rospy.Subscriber('/area_{}/status'.format(area_id), Int8,
-#                                                                 self.area_status_cb, area_id)
-#
-#         # Wait for self.curr_fmeasures to be populated
-#         while set(list(self.curr_fmeasures.keys())) != set(self.areas):
-#             self.debug("Subscribing to area topics...")
-#             rospy.sleep(1)
-#
-#         # Sampled node poses and distance matrix
-#         self.sampled_nodes_poses = self.extract_sampled_node_poses(self.assigned_areas)
-#         self.build_dist_matrix()
-#
-#     def get_assigned_area_index(self, area_id):
-#         """
-#         Gets area index of assigned area, (note area index starts at 1)
-#         :param area_id:
-#         :return:
-#         """
-#         area_idx = self.assigned_areas.index(area_id) + 1
-#         return area_idx
-#
-#     def get_assigned_area_id(self, area_idx):
-#         """
-#         Gets actual assigned area id given area index, (note area index starts at 1)
-#         :param self:
-#         :param area_idx:
-#         :return:
-#         """
-#         area_id = self.assigned_areas[area_idx - 1]
-#         return area_id
-#
-#     def get_current_state(self):
-#         """
-#         Retrieves current state
-#         :return:
-#         """
-#         tlapses = dict()
-#         decay_rates = dict()
-#         for area_idx in self.areas:
-#             area_id = self.get_assigned_area_id(area_idx)
-#             tlapses[area_id] = self.tlapses[area_idx]
-#             decay_rates[area_id] = self.decay_rates_dict[area_idx]
-#
-#         state = (self.sim_t, self.get_assigned_area_id(self.curr_loc_idx), self.battery, tlapses, decay_rates)
-#         return state
-#
-#     # Methods: Run operation
-#     def run_operation(self, filename, freq=1):
-#         """
-#         :return:
-#         """
-#
-#         if self.robot_id < 999:  # Not a dummy robot
-#             rate = rospy.Rate(freq)
-#             rospy.sleep(5)  # Wait for nodes to register
-#
-#             self.register_to_central()
-#
-#             while self.dist_matrix is None:
-#                 self.debug("Initialization: No cluster assignment yet. Waiting for assignment...")
-#                 rospy.sleep(1)
-#
-#             # TODO: To take into account re-assignment of areas
-#
-#             self.sim_t = 0
-#             # while not rospy.is_shutdown() and self.sim_t<self.t_operation:
-#             while not rospy.is_shutdown():
-#                 curr_state = self.get_current_state()
-#                 self.state.append(curr_state)
-#                 self.debug("Curr state: {}".format(curr_state))
-#                 self.robot_status_pub.publish(self.robot_status)
-#                 self.location_pub.publish(self.get_assigned_area_id(self.curr_loc_idx))
-#                 self.status_history.append(self.robot_status)
-#
-#                 # I am learning something new! Each day
-#                 # Freedom comes with courage. And with courage comes freedom.
-#                 # New wave. I create a new culture. A new trend people will appreciate and follow and love
-#                 # A brand new consciousness
-#
-#                 if self.robot_status == robotStatus.IDLE.value:  # Here, robot is available but unassigned
-#                     self.debug('Robot idle')
-#                     self.debug("With dist_matrix: {}".format(bool(self.dist_matrix is not None)))
-#                     if self.dist_matrix is not None:
-#                         self.update_robot_status(robotStatus.READY)
-#
-#                 elif self.robot_status == robotStatus.READY.value:
-#                     self.debug('Robot ready')
-#                     think_start = process_time()
-#                     self.think_decisions()
-#                     think_end = process_time()
-#                     think_elapsed = self.time_elapsed(think_start, think_end)
-#
-#                     bounds = self.strict_bounds()
-#                     self.debug('Strict bounds:{}'.format(bounds))
-#                     self.strict_bounds_list.append(bounds)
-#
-#                     self.process_time_counter.append(think_elapsed)
-#
-#                     self.debug(
-#                         'Best decision: {}. Process time: {}s'.format(self.get_assigned_area_id(self.best_decision_idx),
-#                                                                       think_elapsed))
-#                     self.update_robot_status(robotStatus.IN_MISSION)
-#
-#                 elif self.robot_status == robotStatus.IN_MISSION.value:
-#                     self.debug('Robot in mission. Total distance travelled: {}'.format(self.total_dist_travelled))
-#                     if self.available:
-#                         self.commence_mission()
-#
-#                 elif self.robot_status == robotStatus.CHARGING.value:
-#                     # TODO: Po, we insert self.assign_status = unassigned here for central's reconsideration. HERE!
-#                     self.debug('Waiting for battery to charge up')
-#
-#                 elif self.robot_status == robotStatus.RESTORING_F.value:
-#                     self.debug('Restoring F-measure')
-#
-#                 self.sim_t += 1  # TODO: I think this should be together when elapse the tlapse in each area. After all, this is what the ideal simulation about where there is no idle state for the robot.
-#
-#                 # Update tlapse for each area when all nodes have registered
-#                 if len(self.decisions_made) > 1 or (self.robot_status != robotStatus.IDLE.value) and (
-#                         self.robot_status != robotStatus.READY.value):
-#                     self.update_tlapses_areas(self.sim_t)
-#
-#                 if self.save:
-#                     pu.dump_data(self.strict_bounds_list, '{}_strict_bounds'.format(filename))
-#                     pu.dump_data(self.state, '{}_environment_state'.format(filename))
-#                     pu.dump_data(self.process_time_counter, '{}_robot{}_process_time'.format(filename, self.robot_id))
-#                     pu.dump_data(self.decisions_made, '{}_robot{}_decisions'.format(filename, self.robot_id))
-#                     pu.dump_data((self.decisions_accomplished, self.total_dist_travelled),
-#                                  '{}_robot{}_decisions_acc_travel'.format(filename, self.robot_id))
-#                     pu.dump_data(self.status_history, '{}_robot{}_status_history'.format(filename, self.robot_id))
-#
-#                 rate.sleep()
-#
-#             # #Store results
-#             # self.update_robot_status(robotStatus.SHUTDOWN) #TODO: There should be a notice that come from central actually
-#             # self.robot_status_pub.publish(self.robot_status)
-#             # self.location_pub.publish(self.get_assigned_area_id(self.curr_loc_idx))
-#             # self.status_history.append(self.robot_status)
-#             #
-#             # #Wait before all other nodes have finished dumping their data
-#             # if self.save:
-#             #     pu.dump_data(self.strict_bounds_list, '{}_strict_bounds'.format(filename))
-#             #     pu.dump_data(self.state, '{}_environment_state'.format(filename))
-#             #     pu.dump_data(self.process_time_counter, '{}_robot{}_process_time'.format(filename, self.robot_id))
-#             #     pu.dump_data(self.decisions_made, '{}_robot{}_decisions'.format(filename, self.robot_id))
-#             #     pu.dump_data((self.decisions_accomplished, self.total_dist_travelled), '{}_robot{}_decisions_acc_travel'.format(filename, self.robot_id))
-#             #     pu.dump_data(self.status_history, '{}_robot{}_status_history'.format(filename, self.robot_id))
-#             #     self.debug("Dumped all data.".format(self.robot_id))
-#             # self.shutdown(sleep=10)
-#
-#     """
-#     NOTE: For the subscribed/published topics and data storages, as well as self.debugs! use assigned area ids! All else area index
-#     """
-#
-#     def think_decisions(self):
-#         """
-#         Thinks of the best decision before starting mission
-#         :return:
-#         """
-#         self.best_decision_idx = self.greedy_best_decision()
-#
-#     def time_elapsed(self, think_start, think_end):
-#         """
-#         Process time it took to process
-#         :param think_end:
-#         :return:
-#         """
-#         return think_end - think_start
-#
-#     def commence_mission(self):
-#         """
-#         Commences mission
-#         :return:
-#         """
-#         if self.send2_next_area() == 0:
-#             self.update_robot_status(robotStatus.IDLE)
-#
-#     def send2_next_area(self):
-#         """
-#         Sends the robot to the next area in the optimal path:
-#         :return:
-#         """
-#         if self.best_decision_idx is not None:
-#             self.mission_area_idx = self.best_decision_idx
-#             mission_area_id = self.get_assigned_area_id(self.mission_area_idx)
-#             self.mission_area_idx_pub.publish(mission_area_id)
-#             self.debug('Heading to: {}. {}'.format(mission_area_id, self.sampled_nodes_poses[self.mission_area_idx]))
-#             self.decisions_made.append(mission_area_id)  # store decisions made
-#             self.go_to_target(self.mission_area_idx)
-#             return 1
-#         return 0
-#
-#     def update_robot_status(self, status):
-#         """
-#         Updates robot status
-#         :param status:
-#         :return:
-#         """
-#         self.robot_status = status.value
-#
-#     def distance_travelled_cb(self, msg):
-#         # Updates total distance travelled
-#         # Sets curr robot pose
-#         x, y = msg.pose.pose.position.x, msg.pose.pose.position.y
-#         self.total_dist_travelled += math.dist((self.x, self.y), (x, y))
-#         self.x, self.y = x, y
-#
-#     def battery_level_cb(self, msg):
-#         """
-#         Callback for battery level
-#         :param msg:
-#         :return:
-#         """
-#         self.battery = msg.data
-#
-#     def battery_status_cb(self, msg):
-#         """
-#
-#         :param msg:
-#         :return:
-#         """
-#         self.environment_status[self.charging_station] = msg.data
-#         if msg.data == battStatus.FULLY_CHARGED.value:
-#             if self.robot_id < 999: self.debug("Fully charged!")
-#             self.available = True
-#             self.update_robot_status(robotStatus.IN_MISSION)
-#
-#     def notify_assignment_accomplishment(self, area_id):
-#         """
-#         Assigns clusters to robots
-#         :param clusters:
-#         :return:
-#         """
-#
-#         rospy.wait_for_service("/assignment_accomplishment_server")
-#         try:
-#             notify_server = rospy.ServiceProxy("/cluster_assignment_server", assignmentAccomplishment)
-#             resp = notify_server(self.robot_id, area_id)
-#             self.debug("Robot: {}. Accomplished assignment: {}. Notified server".format(self.robot_id, area_id))
-#         except rospy.ServiceException as e:
-#             rospy.logerr(f"Service call failed: {e}")
-#
-#     def area_status_cb(self, msg, area_id):
-#         """
-#
-#         :param msg:
-#         :return:
-#         """
-#         area_idx = self.get_assigned_area_index(area_id)
-#         status = msg.data
-#         self.environment_status[area_idx + 1] = int(status)
-#
-#         if msg.data == areaStatus.RESTORED_F.value:
-#             if self.robot_id < 999: self.debug("Area {} fully restored!".format(area_id))
-#             self.tlapses[area_idx] = 0  # Reset the tlapse since last restored for the newly restored area
-#
-#             self.notify_assignment_accomplishment(area_id)  # Notifies central that recent assignment is accomplished
-#
-#             self.available = True
-#             self.update_robot_status(robotStatus.IN_MISSION)
-#
-#     def decay_rate_cb(self, msg, area_id):
-#         """
-#         Store decay rate
-#         :param msg:
-#         :param area_id:
-#         :return:
-#         """
-#         area_idx = self.get_assigned_area_index(area_id)
-#         if self.assigned_areas is not None and self.decay_rates_dict[area_idx] == None:
-#             if self.robot_id < 999: self.debug("Area {} decay rate: {}".format(area_id, msg.data))
-#             self.decay_rates_dict[area_idx] = msg.data
-#
-#     def area_fmeasure_cb(self, msg, area_id):
-#         """
-#         Updates fmeasure of area
-#         :param msg:
-#         :param area_id:
-#         :return:
-#         """
-#         area_idx = self.get_assigned_area_index(area_id)
-#         self.curr_fmeasures[area_idx] = msg.data
-#
-#     def debug(self, msg):
-#         pu.log_msg('robot', self.robot_id, msg, self.debug_mode)
-#
-#     def shutdown(self, sleep):
-#         self.debug("Reached {} time operation. Shutting down...".format(self.t_operation))
-#         kill_nodes(sleep)
-#
-#
-# if __name__ == '__main__':
-#     # os.chdir('/home/ameldocena/.ros/int_preservation/results')
-#     os.chdir('/root/catkin_ws/src/results/int_preservation')
-#     filename = rospy.get_param('/file_data_dump')
-#     Robot('heuristic_decision').run_operation(filename)
