--- scripts/online_posterior_sampling_v1.py
@@ -1,715 +1,0 @@
-#!/usr/bin/env python
-
-"""
-Online Posterior Sampling for POMDP under STAR settings
-
-Papers:
-1. Posterior Sampling-based Online Learning for Episodic POMDPs, Tang, et al., 2024. https://doi.org/10.48550/arXiv.2310.10107
-2. (More) Efficient Reinforcement Learning via Posterior Sampling, Osband, et al., NIPS 2013. https://papers.nips.cc/paper_files/paper/2013/file/6a5889bb0190d0211a991f47bb19a777-Paper.pdf
-"""
-import rospy
-import actionlib
-import json
-import math
-from time import process_time
-import pickle
-import numpy as np
-from loss_fcns import *
-from pruning import *
-import project_utils as pu
-from nav_msgs.msg import Odometry
-from std_msgs.msg import Int8, Float32
-from move_base_msgs.msg import MoveBaseAction, MoveBaseGoal
-from nav_msgs.srv import GetPlan
-from int_preservation.srv import flevel, flevelRequest
-from status import areaStatus, battStatus, robotStatus
-from reset_simulation import *
-
-def request_fmeasure(area, msg=True):
-    """
-    Service request for F-measure
-    :param msg:
-    :return: (tlapse, F) #tlapse here is time lapsed since decay model of area has been updated
-    """
-    rospy.wait_for_service("/flevel_server_" + str(area))
-    flevel_service = rospy.ServiceProxy("/flevel_server_" + str(area), flevel)
-    request = flevelRequest(msg)
-    result = flevel_service(request)
-    return result.fmeasure
-
-
-INDEX_FOR_X = 0
-INDEX_FOR_Y = 1
-SUCCEEDED = 3  # GoalStatus ID for succeeded, http://docs.ros.org/en/api/actionlib_msgs/html/msg/GoalStatus.html
-SHUTDOWN_CODE = 99
-
-
-class Robot:
-    def __init__(self, node_name):
-        """
-
-        :param node_name:
-        :param areas:
-        :param est_distance_matrix:
-        :param est_batt_consumption_matrix:
-        """
-
-        rospy.init_node(node_name, anonymous=True)
-
-        # Parameters
-        self.robot_id = rospy.get_param("~robot_id")
-        self.debug_mode = rospy.get_param("/debug_mode")
-        self.robot_velocity = rospy.get_param("/robot_velocity")  # Linear velocity of robot; we assume linear and angular are relatively equal
-        self.gamma = rospy.get_param("/gamma")  # discount factor
-        self.max_fmeasure = rospy.get_param("/max_fmeasure")  # Max F-measure of an area
-        self.max_battery = rospy.get_param("/max_battery")  # Max battery
-        self.battery_reserve = rospy.get_param("/battery_reserve")  # Battery reserve
-
-        f_thresh = rospy.get_param("/f_thresh")
-        self.fsafe, self.fcrit = f_thresh
-
-        batt_consumed_per_time = rospy.get_param("/batt_consumed_per_time")
-        self.batt_consumed_per_travel_time, self.batt_consumed_per_restored_f = batt_consumed_per_time
-
-        self.dec_steps = rospy.get_param("/dec_steps")  # STAR
-        self.restoration = rospy.get_param("/restoration")
-        self.noise = rospy.get_param("/noise")
-        self.nareas = rospy.get_param("/nareas")  # Sample nodes from voronoi equal to area count #STAR
-        self.areas = [int(i + 1) for i in range(self.nareas)]  # list of int area IDs
-        self.tolerance = rospy.get_param("/move_base_tolerance")
-        self.t_operation = rospy.get_param("/t_operation")  # total duration of the operation
-        self.save = rospy.get_param("/save")  # Whether to save data
-
-        # Initialize variables
-        charging_station_coords = rospy.get_param("~initial_pose_x"), rospy.get_param("~initial_pose_y")  # rospy.get_param("/charging_station_coords")
-        charging_pose_stamped = pu.convert_coords_to_PoseStamped(charging_station_coords)
-        self.sampled_nodes_poses = [charging_pose_stamped]  # list container for sampled nodes of type PoseStamped
-
-        # Pickle load the sampled area poses
-        with open('{}.pkl'.format(rospy.get_param("/file_sampled_areas")), 'rb') as f:
-            sampled_areas_coords = pickle.load(f)
-        for area_coords in sampled_areas_coords['n{}_p{}'.format(self.nareas, rospy.get_param("/placement"))]:
-            pose_stamped = pu.convert_coords_to_PoseStamped(area_coords)
-            self.sampled_nodes_poses.append(pose_stamped)
-
-        self.x, self.y = 0.0, 0.0  # Initialize robot pose
-        self.charging_station = 0
-        self.curr_loc = self.charging_station  # Initial location robot is the charging station
-        self.battery = self.max_battery  # Initialize battery at max, then gets updated by subscribed battery topic
-        self.best_decision = None
-        self.dist_matrix = None
-        self.mission_area = None
-        self.robot_status = robotStatus.IDLE.value
-        self.available = True
-        self.curr_fmeasures = dict()  # container of current F-measure of areas
-        self.decay_rates_dict = dict()  # dictionary for decay rates
-        self.recorded_fdata = dict()  # dictionary of recorded data collected during mission per area
-        self.tlapses = dict()  # dictionary containing tlapse for each area since last restoration
-
-        for area in self.areas:
-            self.curr_fmeasures[area] = self.max_fmeasure
-            self.decay_rates_dict[area] = None
-            self.recorded_fdata[area] = list()
-            self.tlapses[area] = 0
-
-        self.decay_rates_counter = 0  # counter for stored decay rates; should be equal to number of areas
-        self.decisions_made, self.decisions_accomplished, self.status_history = [], [], []  # record of data
-        self.total_dist_travelled = 0  # total distance travelled
-        self.process_time_counter = []  # container for time it took to come up with decision
-
-        #Variables for online PSRL
-        self.inference = rospy.get_param("/inference") #Whether oracle or None; toggle CONSIDER_REPLAN which updates decay belief by posterior sampling
-        self.observations_dict = dict()
-        self.posterior_belief_disbn = dict()
-        init_mean, init_sd = 0.001875, 0.000875 #Prior data/mission (params are based on actual experimental settings)
-        for area in self.areas:
-            self.observations_dict[area] = list()
-            self.posterior_belief_disbn[area] = (init_mean, init_sd)
-
-        # We sum this up
-        self.environment_status = dict()
-        for node in range(self.nareas + 1):
-            self.environment_status[node] = 999
-
-        # Publishers/Subscribers
-        rospy.Subscriber('/robot_{}/odom'.format(self.robot_id), Odometry, self.distance_travelled_cb, queue_size=1)
-
-        # Service request to move_base to get plan : make_Plan
-        server = '/robot_' + str(self.robot_id) + '/move_base_node/make_plan'
-        rospy.wait_for_service(server)
-        self.get_plan_service = rospy.ServiceProxy(server, GetPlan)
-        # self.debug("Getplan service: {}".format(self.get_plan_service))
-
-        rospy.Subscriber('/robot_{}/battery_status'.format(self.robot_id), Int8, self.battery_status_cb)
-        rospy.Subscriber('/robot_{}/battery'.format(self.robot_id), Float32, self.battery_level_cb)
-
-        for area in self.areas:
-            rospy.Subscriber('/area_{}/decay_rate'.format(area), Float32, self.decay_rate_cb, area)
-            rospy.Subscriber('/area_{}/fmeasure'.format(area), Float32, self.area_fmeasure_cb, area)  # REMARK: Here we assume that we have live measurements of the F-measures
-            rospy.Subscriber('/area_{}/status'.format(area), Int8, self.area_status_cb, area)
-
-        self.robot_status_pub = rospy.Publisher('/robot_{}/robot_status'.format(self.robot_id), Int8, queue_size=1)
-        self.mission_area_pub = rospy.Publisher('/robot_{}/mission_area'.format(self.robot_id), Int8, queue_size=1)
-
-        # Action client to move_base
-        self.robot_goal_client = actionlib.SimpleActionClient('/robot_' + str(self.robot_id) + '/move_base',
-                                                              MoveBaseAction)
-        self.robot_goal_client.wait_for_server()
-
-
-    # METHODS: Node poses and distance matrix
-    def get_plan_request(self, start_pose, goal_pose, tolerance):
-        """
-        Sends a request to GetPlan service to create a plan for path from start to goal without actually moving the robot
-        :param start_pose:
-        :param goal_pose:
-        :param tolerance:
-        :return:
-        """
-        req = GetPlan()
-        req.start = start_pose
-        req.goal = goal_pose
-        req.tolerance = tolerance
-        server = self.get_plan_service
-        result = server(req.start, req.goal, req.tolerance)
-        path = result.plan.poses
-        return path
-
-    def decouple_path_poses(self, path):
-        """
-        Decouples a path of PoseStamped poses; returning a list of x,y poses
-        :param path: list of PoseStamped
-        :return:
-        """
-        list_poses = list()
-        for p in path:
-            x, y = p.pose.position.x, p.pose.position.y
-            list_poses.append((x, y))
-        return list_poses
-
-    def compute_path_total_dist(self, list_poses):
-        """
-        Computes the total path distance
-        :return:
-        """
-        total_dist = 0
-        for i in range(len(list_poses) - 1):
-            dist = math.dist(list_poses[i], list_poses[i + 1])
-            total_dist += dist
-        return total_dist
-
-    def compute_dist_bet_areas(self, area_i, area_j, tolerance):
-        """
-        Computes the distance between area_i and area_j:
-            1. Call the path planner between area_i and area_j
-            2. Decouple the elements of path planning
-            3. Compute the distance then total distance
-        :param area_i: PoseStamped
-        :param area_j: PoseStamped
-        :return:
-        """
-        path = self.get_plan_request(area_i, area_j, tolerance)
-        list_poses = self.decouple_path_poses(path)
-        total_dist = self.compute_path_total_dist(list_poses)
-        return total_dist
-
-    def build_dist_matrix(self):
-        """
-        Builds the distance matrix among areas
-        :return:
-        """
-        n = len(self.sampled_nodes_poses)
-        self.dist_matrix = np.zeros((n, n))
-
-        for i in range(n):
-            for j in range(n):
-                area_i, area_j = self.sampled_nodes_poses[i], self.sampled_nodes_poses[j]
-                if area_i != area_j:
-                    dist = self.compute_dist_bet_areas(area_i, area_j, self.tolerance)
-                    self.dist_matrix[i, j] = dist
-
-        self.debug("Dist matrix: {}".format(self.dist_matrix))
-
-    # METHODS: Send robot to area
-    def go_to_target(self, goal_idx):
-        """
-        Action client to move_base to move to target goal
-        Goal is PoseStamped msg
-        :param: goal_idx, index of goal in sampled_nodes_poses list
-        :return:
-        """
-        goal = self.sampled_nodes_poses[goal_idx]
-        self.send_robot_goal(goal)
-
-    def send_robot_goal(self, goal):
-        """
-        Sends robot to goal via action client
-        :param robot:
-        :param goal: PoseStamped object
-        :return:
-        """
-        movebase_goal = MoveBaseGoal()
-        movebase_goal.target_pose = goal
-        self.available = False
-        action_goal_cb = (lambda state, result: self.action_send_done_cb(state, result, self.robot_id))
-        self.robot_goal_client.send_goal(movebase_goal, done_cb=action_goal_cb, active_cb=self.action_send_active_cb)
-
-    def action_send_active_cb(self):
-        """
-        Sets robot as unavailable when pursuing goal
-        :return:
-        """
-        self.available = False
-        self.update_robot_status(robotStatus.IN_MISSION)
-
-    def action_send_done_cb(self, state, result, robot_id):
-        """
-        What happens after robot succeeded in fulfilling assigned action (recharge or restore area)
-        :param msg:
-        :return:
-        """
-        if state == SUCCEEDED:
-            self.debug("Arrived at mision_area: {}".format(self.mission_area))
-            self.curr_loc = self.mission_area
-            self.decisions_accomplished.append(self.mission_area)
-            self.best_decision = None
-            if self.mission_area == self.charging_station:
-                self.update_robot_status(robotStatus.CHARGING)
-            else:
-                data = request_fmeasure(self.curr_loc)
-                measured_f = float(data)
-                self.recorded_fdata[self.curr_loc].append(measured_f)
-                observed_decay_rate = get_decay_rate(self.max_fmeasure, measured_f, self.tlapses[self.curr_loc])
-                self.observations_dict[self.curr_loc].append(observed_decay_rate)
-
-                self.debug("Prior belief decay distribution: {}".format(self.posterior_belief_disbn[self.curr_loc]))
-                self.bayesian_belief_update(self.curr_loc)
-                self.debug("Observed decay rate: {}. Decay rate history: {}. Updated posterior decay distribution: {}".format(observed_decay_rate, self.observations_dict[self.curr_loc], self.posterior_belief_disbn[self.curr_loc]))
-                self.update_robot_status(robotStatus.RESTORING_F)
-
-    def mean_duration_decay(self, duration_matrix, area):
-        """
-        Measures the average duration that an area decays (when robot commits to another decision than restoring area)
-        Note that each column of the duration matrix represents that the corresponding area is the decision that the robot is committing to.
-        We thus delete the column of the area, and then take the average to measure the average duration.
-        :param duration_matrix:
-        :param area:
-        :return:
-        """
-        submatrix = np.delete(duration_matrix, area, axis=1)
-        return np.mean(submatrix)
-
-    def estimate_battery_params(self, decision, curr_battery, curr_loc, fmeasures, noise):
-        """
-        Measures battery consumption and feasible battery
-        :param decision:
-        :return:
-        """
-        # Battery consumed travel and preserve area (if not charging station)
-        battery_consumption = self.consume_battery(start_area=curr_loc, next_area=decision,
-                                                   curr_measure=fmeasures[decision],
-                                                   noise=noise)
-        # Battery consumed travel back to charging station
-        battery_consumption_backto_charging_station = self.consume_battery(start_area=decision,
-                                                                           next_area=self.charging_station,
-                                                                           curr_measure=None,
-                                                                           noise=noise)
-        total_battery_consumption = battery_consumption + battery_consumption_backto_charging_station
-        # Feasible batt = current batt - (consumption of decision + consumption to travel back to charging station)
-        feasible_battery = curr_battery - total_battery_consumption
-
-        return total_battery_consumption, feasible_battery
-
-    def compute_reward(self, fmeasure):
-        """
-        Computes reward of action based on loss that has been saved up
-        :param area:
-        :param duration:
-        :return:
-        """
-        reward = -loss_fcn(self.max_fmeasure, fmeasure)
-        return reward
-
-    def policy_max_value(self):
-        """
-        Policy that takes the feasible action that has the max value
-        :return:
-        """
-        # Evaluate feasible decisions
-        decision_array = []
-        for decision in self.areas:
-            # Battery consumption
-            battery_consumption, feasible_battery = self.estimate_battery_params(decision, self.battery, self.curr_loc,
-                                                                                 self.curr_fmeasures, self.noise)
-            # self.debug("Batt consumption: {}. Feasible batt: {}".format(battery_consumption, feasible_battery))
-
-            if not prune(self.battery, battery_consumption, self.battery_reserve) and decision != self.curr_loc:
-                #Compute the value based on belief decay rate`
-                duration = self.compute_duration(self.curr_loc, decision, self.curr_fmeasures[decision],
-                                                 self.restoration, self.noise)
-                updated_fmeasures = self.adjust_fmeasures(self.curr_fmeasures.copy(), decision, duration)  # F-measure of areas adjusted accordingly, i.e., consequence of decision
-                value = self.compute_reward(updated_fmeasures[decision])
-                decision_array.append((decision, value, feasible_battery))
-
-        best_decision = self.charging_station
-
-        if len(decision_array) > 0:
-            best_decision = self.get_max_value(decision_array)
-
-        return best_decision
-
-    def compute_duration(self, start_area, next_area, curr_measure, restoration, noise):
-        """
-        Computes (time) duration of operation, which includes travelling distance plus restoration, if any
-        :param distance:
-        :param restoration: restore a measure (if not None) back to full measure per second
-        :param noise: expected noise in distance travelled
-        :return:
-        """
-
-        # Travel distance
-        distance = self.dist_matrix[int(start_area), int(next_area)]
-        distance += noise * distance  # distance + noise
-        time = (distance / self.robot_velocity)
-
-        # If next area is not the charging station: the restoration is the f-measure; else, the restoration is the battery level
-        if next_area != self.charging_station:
-            max_restore = self.max_fmeasure
-        else:
-            max_restore = self.max_battery
-
-        # Restoration time: If there is need for restoration
-        if (curr_measure is not None) and (restoration is not None):
-            restore_time = (max_restore - curr_measure) / restoration
-            time += restore_time
-
-        return time
-
-    def consume_battery(self, start_area, next_area, curr_measure, noise):
-        """
-        Consumes curr_battery for the duration of the operation.
-        This duration includes the distance plus F-measure restoration, if any
-        :param curr_battery:
-        :param duration:
-        :return:
-        """
-
-        # Batt consumed in travel
-        distance = self.dist_matrix[int(start_area), int(next_area)]
-        distance += noise * distance
-        travel_time = (distance / self.robot_velocity)
-        battery_consumed = self.batt_consumed_per_travel_time * travel_time
-
-        if next_area != self.charging_station:
-            battery_consumed += self.batt_consumed_per_restored_f * (self.max_fmeasure - curr_measure)
-
-        return battery_consumed
-
-    def adjust_fmeasures(self, fmeasures, visit_area, duration):
-        """
-        Adjusts the F-measures of all areas in robot's mind. The visit area will be restored to max, while the other areas will decay for
-        t duration. Note that the charging station is not part of the areas to monitor. And so, if the visit_area is the
-        charging station, then all of the areas will decay as duration passes by.
-        :param fmeasures:
-        :param visit_area:
-        :param t:
-        :return:
-        """
-        for area in self.areas:
-            if area == visit_area:
-                fmeasures[area] = self.max_fmeasure
-            else:
-                tlapse_decay = get_time_given_decay(self.max_fmeasure, fmeasures[area],
-                                                    self.decay_rates_dict[area]) + duration
-                fmeasures[area] = decay(self.decay_rates_dict[area], tlapse_decay, self.max_fmeasure)
-
-        return fmeasures
-
-    def compute_net_loss(self, fmeasures):
-        """
-        Computes the net loss, (i.e., the sum of losses) of the fmeasures, which is a consequence of a decision
-        Steps:
-            1. Computes the loss for each of the F-measure of the areas
-            2. Sums up the losses to get the cost of the decision
-        :param fmeasures:
-        :return:
-        """
-        netloss = compute_cost_fmeasures(fmeasures, self.fsafe, self.fcrit)
-        return netloss
-
-    def get_max_value(self, dec_arr):
-        """
-        Returns the best decision in an array by sorting values descendingly first then by remaining feasible battery.
-        :param tree:
-        :return:
-        """
-        # Sort the decisions: the cost is key while the value is decision
-        sorted_decisions = sorted(dec_arr, key=lambda x: (-x[-2], -x[-1]))
-        self.debug("Decisions sorted by value, feasible batt: {}".format(sorted_decisions))
-        self.debug("Best decision (branch info): {}".format(sorted_decisions[0]))
-        best_decision = sorted_decisions[0][0]  # pick the decision with least net loss and most available feasible battery
-        return best_decision
-
-    def update_tlapses_areas(self):
-        """
-        Lapses all time elapsed for each area
-        :return:
-        """
-        for area in self.areas:
-            self.tlapses[area] += 1
-        self.debug("Time elapsed since last restored: {}".format(self.tlapses))
-
-    def bayesian_belief_update(self, area):
-        """
-        Performs bayesian update on belief decay distribution for area
-
-        Note: Design assumes we have a new observation for
-
-        :return:
-        """
-        #Prior
-        prior_mean, prior_sd = self.posterior_belief_disbn[area] #list of current posterior belief per area, now set as prior
-
-        #Empirical mean and sd
-        n = len(self.observations_dict[area])
-        mean_data = np.mean(self.observations_dict[area]) #list of observations per area
-        sd_data = np.std(self.observations_dict[area], ddof=1)
-
-        #Bayesian udpate
-        sd_posterior_sq = 1 / (1/prior_sd**2 + n/sd_data**2)
-        mean_posterior = ((prior_mean / prior_sd**2) + ((n*mean_data) / sd_data**2)) * sd_posterior_sq
-        sd_posterior = np.sqrt(sd_posterior_sq)
-
-        #Update belief
-        self.posterior_belief_disbn[area] = (mean_posterior, sd_posterior)
-
-    def sample_posterior_belief(self, area):
-        """
-        Samples decay rate from current posterior belief on decay distribution
-        :return:
-        """
-
-        mean_belief, sd_belief = self.posterior_belief_disbn[area]
-        decay_rate = np.random.normal(loc=mean_belief, scale=sd_belief)
-        return decay_rate
-
-    def run_operation(self, filename, freq=1):
-        """
-        :return:
-        """
-
-        if self.robot_id < 999:
-            rate = rospy.Rate(freq)
-
-            while len(self.sampled_nodes_poses) != self.nareas + 1:
-                self.debug("Insufficient data. Sampled nodes poses: {}/{}".format(len(self.sampled_nodes_poses), self.nareas + 1))
-                rate.sleep()  # Data for decay rates haven't registered yet
-
-            for area in self.areas:
-                belief_decay = self.sample_posterior_belief(area) #Sample posterior belief based on prior knowledge
-                self.decay_rates_dict = belief_decay
-                self.observations_dict[area].append(belief_decay) #Append to history of observed decay, (set as initial)
-
-            self.debug("Sufficent data. Decay rates: {}. Sampled nodes poses: {}".format(self.decay_rates_dict,
-                                                                                         self.sampled_nodes_poses))  # Prior knowledge of decay rates
-
-            self.build_dist_matrix()
-            t = 0
-            while not rospy.is_shutdown() and t < self.t_operation:
-                self.robot_status_pub.publish(self.robot_status)
-                self.status_history.append(self.robot_status)
-                if self.robot_status == robotStatus.IDLE.value:
-                    self.debug('Robot idle')
-                    if self.dist_matrix is not None:
-                        self.update_robot_status(robotStatus.READY)
-
-                elif self.robot_status == robotStatus.READY.value:
-                    self.debug('Robot ready')
-                    think_start = process_time()
-                    self.think_decisions()
-                    think_end = process_time()
-                    think_elapsed = self.time_elapsed(think_start, think_end)
-                    self.process_time_counter.append(think_elapsed)
-                    self.debug('Best decision: {}. Process time: {}s'.format(self.best_decision, think_elapsed))
-                    self.update_robot_status(robotStatus.IN_MISSION)
-
-                elif self.robot_status == robotStatus.IN_MISSION.value:
-                    self.debug('Robot in mission. Total distance travelled: {}'.format(self.total_dist_travelled))
-                    if self.available:
-                        self.commence_mission()
-
-                elif self.robot_status == robotStatus.CHARGING.value:
-                    self.debug('Waiting for battery to charge up')
-
-                elif self.robot_status == robotStatus.RESTORING_F.value:
-                    self.debug('Restoring F-measure')
-
-                elif self.robot_status == robotStatus.CONSIDER_REPLAN.value:
-                    self.debug('Consider re-plan...')
-                    #Posterior sampling of decay rates based on belief
-                    for area in self.areas:
-                        belief_decay = self.sample_posterior_belief(area)
-                        self.decay_rates_dict = belief_decay
-                    self.debug("Sampled posterior decay rates: {}".format(self.decay_rates_dict))
-                    self.update_robot_status(robotStatus.IN_MISSION)
-
-                if len(self.decisions_made)>0 or (self.robot_status != robotStatus.IDLE.value) and (
-                        self.robot_status != robotStatus.READY.value) and (
-                        self.robot_status != robotStatus.CONSIDER_REPLAN.value):
-                    self.update_tlapses_areas()  # Update the tlapse per area
-                    self.compute_curr_fmeasures()
-                t += 1
-                rate.sleep()
-
-            # Store results
-            self.update_robot_status(robotStatus.SHUTDOWN)
-            self.robot_status_pub.publish(self.robot_status)
-            self.status_history.append(self.robot_status)
-
-            # Wait before all other nodes have finished dumping their data
-            if self.save:
-                if self.inference is not None:
-                    pu.dump_data((self.recorded_fdata, self.observations_dict),
-                                 '{}_robot{}_recorded_data'.format(filename, self.robot_id))
-
-                pu.dump_data(self.process_time_counter, '{}_robot{}_process_time'.format(filename, self.robot_id))
-                pu.dump_data(self.decisions_made, '{}_robot{}_decisions'.format(filename, self.robot_id))
-                pu.dump_data((self.decisions_accomplished, self.total_dist_travelled),
-                             '{}_robot{}_decisions_acc_travel'.format(filename, self.robot_id))
-                pu.dump_data(self.status_history, '{}_robot{}_status_history'.format(filename, self.robot_id))
-                self.debug("Dumped all data.".format(self.robot_id))
-            self.shutdown(sleep=10)
-
-    def think_decisions(self):
-        """
-        Thinks of the best decision before starting mission
-        :return:
-        """
-        self.best_decision = self.policy_max_value()
-
-    def time_elapsed(self, think_start, think_end):
-        """
-        Process time it took to process
-        :param think_end:
-        :return:
-        """
-        return think_end - think_start
-
-    def commence_mission(self):
-        """
-        Commences mission
-        :return:
-        """
-        if self.send2_next_area() == 0:
-            self.update_robot_status(robotStatus.IDLE)
-
-    def send2_next_area(self):
-        """
-        Sends the robot to the next area in the optimal path/decision:
-        :return:
-        """
-        if self.best_decision is not None:
-            self.mission_area = self.best_decision
-            self.mission_area_pub.publish(self.mission_area)
-            self.debug('Heading to: {}. {}'.format(self.mission_area, self.sampled_nodes_poses[self.mission_area]))
-            self.decisions_made.append(self.mission_area)  # store decisions made
-            self.go_to_target(self.mission_area)
-            return 1
-        return 0
-
-    def update_robot_status(self, status):
-        """
-        Updates robot status
-        :param status:
-        :return:
-        """
-        self.robot_status = status.value
-
-    def distance_travelled_cb(self, msg):
-        # Updates total distance travelled
-        # Sets curr robot pose
-        x, y = msg.pose.pose.position.x, msg.pose.pose.position.y
-        self.total_dist_travelled += math.dist((self.x, self.y), (x, y))
-        self.x, self.y = x, y
-
-    def battery_level_cb(self, msg):
-        """
-        Callback for battery level
-        :param msg:
-        :return:
-        """
-        self.battery = msg.data
-
-    def battery_status_cb(self, msg):
-        """
-
-        :param msg:
-        :return:
-        """
-        self.environment_status[self.charging_station] = msg.data
-        if msg.data == battStatus.FULLY_CHARGED.value:
-            if self.robot_id < 999: self.debug("Fully charged!")
-            self.available = True
-            self.update_robot_status(robotStatus.IN_MISSION)
-
-    def area_status_cb(self, msg, area_id):
-        """
-
-        :param msg:
-        :return:
-        """
-        self.environment_status[area_id] = msg.data
-        if msg.data == areaStatus.RESTORED_F.value:
-            self.available = True
-            self.tlapses[area_id] = 0
-            if self.robot_id < 999: self.debug("Area {} fully restored! tlapse reset...".format(area_id))
-
-            if (self.inference is not None) and (self.inference != 'oracle'):
-                self.update_robot_status(robotStatus.CONSIDER_REPLAN)
-            else:
-                self.update_robot_status(robotStatus.IN_MISSION)
-
-    def decay_rate_cb(self, msg, area_id):
-        """
-        Store decay rate
-        :param msg:
-        :param area_id:
-        :return:
-        """
-        if self.inference == 'oracle':
-            if self.decay_rates_dict[area_id] != msg.data: self.debug(
-                "Oracle knowledge, change in decay in area {}: {}".format(area_id, msg.data))
-            self.decay_rates_dict[area_id] = msg.data  # A subscribed topic. Oracle knows exactly the decay rate happening in area
-
-    def area_fmeasure_cb(self, msg, area_id):
-        """
-        Updates fmeasure of area
-        :param msg:
-        :param area_id:
-        :return:
-        """
-        if self.inference == 'oracle':
-            self.curr_fmeasures[area_id] = msg.data
-
-    def compute_curr_fmeasures(self):
-        """
-        Computes current fmeasures based on tlapse and decay rates
-        :return:
-        """
-        for area in self.areas:
-           self.curr_fmeasures[area] = decay(self.decay_rates_dict[area], self.tlapses[area], self.max_fmeasure)
-        # self.debug("Used for computation. Tlapses: {}. Decay rates: {}".format(self.tlapses, self.decay_rates_dict))
-        # self.debug("Computed current f-measures: {}".format(self.curr_fmeasures))
-
-    def debug(self, msg):
-        pu.log_msg('robot', self.robot_id, msg, self.debug_mode)
-
-    def shutdown(self, sleep):
-        self.debug("Reached {} time operation. Shutting down...".format(self.t_operation))
-        kill_nodes(sleep)
-
-
-if __name__ == '__main__':
-    os.chdir('/root/catkin_ws/src/results/int_preservation')
-    filename = rospy.get_param('/file_data_dump')
-    Robot('online_posterior_sampling').run_operation(filename)
