+++ scripts/run_experiment.py
@@ -246,9 +246,9 @@ if __name__ == '__main__':
     # run_experiment('heuristic_decision', 'office', 12, placement, 'non_uniform', 2100, nrobots=2,
     #                inference='oracle', dec_steps=1, discount=0.00, exploration=0.0, ntrials=(0, 3), task_scheduler='central_planner', save=True)
     #
-    # run_experiment('heuristic_decision', 'office', 12, placement, 'non_uniform', 2100, nrobots=2,
-    #                inference='oracle', dec_steps=4, discount=0.75, exploration=0.0, ntrials=(0, 3),
-    #                task_scheduler='central_planner', save=True)
+    run_experiment('online_posterior_sampling', 'office', 12, placement, 'non_uniform', 2100, nrobots=1,
+                   inference=None, dec_steps=4, discount=0.75, exploration=0.0, ntrials=(0, 1),
+                   task_scheduler='central_planner', save=False)
     #
     # run_experiment('heuristic_decision', 'office', 12, placement, 'non_uniform', 2100, nrobots=2,
     #                inference='oracle', dec_steps=6, discount=0.75, exploration=0.0, ntrials=(0, 3), task_scheduler='central_planner', save=True)
@@ -257,13 +257,13 @@ if __name__ == '__main__':
     # run_experiment('heuristic_decision', 'office', 4, placement, 'non_uniform', 50,
     #                nrobots=2, inference='oracle', dec_steps=6, discount=0.75, exploration=0.0, ntrials=(0, 1), task_scheduler='central_planner', save=False) #3100
 
-    run_experiment('dynamic_programming', 'office', 12, placement, 'non_uniform', 2100,
-                   nrobots=2, inference='oracle', dec_steps=4, ntrials=(0, 3),
-                   task_scheduler='central_planner', save=True)  # 3100
-
-    run_experiment('dynamic_programming', 'office', 12, placement, 'non_uniform', 2100,
-                   nrobots=2, inference='oracle', dec_steps=6, ntrials=(0, 3),
-                   task_scheduler='central_planner', save=True)  # 3100
+    # run_experiment('dynamic_programming', 'office', 12, placement, 'non_uniform', 2100,
+    #                nrobots=2, inference='oracle', dec_steps=4, ntrials=(0, 3),
+    #                task_scheduler='central_planner', save=True)  # 3100
+    #
+    # run_experiment('dynamic_programming', 'office', 12, placement, 'non_uniform', 2100,
+    #                nrobots=2, inference='oracle', dec_steps=6, ntrials=(0, 3),
+    #                task_scheduler='central_planner', save=True)  # 3100
 
     #
     # run_experiment('rma_search', 'cluttered', 12, placement, 'non_uniform', 2100,
